### 2023-11-17 #150 @billbrod
Plenoptic
Submitting Author: Billy Broderick (@billbrod)
All current maintainers: @billbrod (there are several other authors / contributors, but I'm the only committed to maintaining)
Package Name: Plenoptic
One-Line Description of Package: a python library for model-based synthesis of perceptual stimuli
Repository Link:  https://github.com/plenoptic-org/plenoptic
Version submitted: v1.0.2
Reviewers assigned: 2023-12-13
Reviews due: 2024-01-19
Version accepted: TBD

---

- Include a brief paragraph describing what your package does: `plenoptic` provides tools to help researchers understand their model by synthesizing novel informative stimuli (e.g., images, movies, or sound clips, depending on the model), which help build intuition for what features the model ignores and what it is sensitive to.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

I am not sure which of the above categories I fit under, but the discussion in my pre-submission inquiry (#97) assured me that I'm in scope.

Researchers in vision science, neuroscience, and machine learning. The goal is to generate novel stimuli (images, videos, audio) that researchers can then use in new experiments to better understand their computational models.

Not that I'm aware of. There are several bits of related research code, but no packages. Research code includes: https://github.com/ArturoDeza/Fast-Texforms, https://github.com/brialorelle/texformgen, https://github.com/ProgramofComputerGraphics/PooledStatisticsMetamers, https://github.com/freeman-lab/metamers/, among others.

#97 , @NickleDave

Does the JOSS and pyOpenSci review happen simultaneously? I was going to add the `paper.md` and submit to JOSS after going through this review.
