### 2023-11-28 d28d265
adds synthesis_video.py
this script contains the code to generate the synthesis schematic video,
though it's not actually run
