+++ docs/scripts/synthesis_video.py
@@ -1,0 +1,150 @@
+#!/usr/bin/env python3
+
+import einops
+import matplotlib.pyplot as plt
+import plenoptic as po
+import matplotlib as mpl
+from matplotlib import animation
+
+def animate_schematic(metamer, fig, axes_idx, framerate=10,
+                      pixel_value_subsample=.01):
+    """Animate synthesis_schematic figure.
+
+    Parameters
+    ----------
+    metamer : po.Metamer
+        The Metamer object to grab data from
+    iteration : int or None, optional
+        Which iteration to display. If None, we show the most recent one.
+        Negative values are also allowed.
+    pixel_value_subsample : float, optional
+        Float between 0 and 1 giving the percentage of image pixels to plot.
+        Done to make visualization clearer.
+    kwargs :
+        passed to metamer.plot_synthesis_status
+
+    Returns
+    -------
+    anim : matplotlib.animation.FuncAnimation
+        The animation object. In order to view, must convert to HTML
+        or save.
+    """
+    def movie_plot(i):
+        artists = []
+        artists.extend(po.tools.update_plot(fig.axes[axes_idx['display_metamer']],
+                                            data=metamer.saved_metamer[i]))
+        img_vals = []
+        for img in [metamer.image, metamer.saved_metamer[i]]:
+            vals = img.flatten()[::int(1/pixel_value_subsample)]
+            img_vals.append(vals.detach())
+        artists.extend(po.tools.update_plot(fig.axes[axes_idx['plot_pixel_values']],
+                                            data=einops.rearrange(img_vals, 'a b -> 1 1 b a')))
+        rep_vals = []
+        for vals in [metamer.target_representation, metamer.model(metamer.saved_metamer[i])]:
+            rep_vals.append(vals.flatten().detach())
+        artists.extend(po.tools.update_plot(fig.axes[axes_idx['plot_representation_comparison']],
+                                            data=einops.rearrange(rep_vals, 'a b -> 1 1 b a')))
+        return artists
+
+    anim = animation.FuncAnimation(fig, movie_plot, frames=len(metamer.saved_metamer),
+                                   blit=True, interval=1000./framerate, repeat=False)
+    plt.close(fig)
+    return anim
+
+
+def synthesis_schematic(metamer, iteration=0, pixel_value_subsample=.01):
+    """Create schematic of synthesis, for animating.
+
+    WARNING: Currently, only works with images of size (256, 256), will need a
+    small amount of tweaking to work with differently sized images. (And may
+    never look quite as good with them)
+
+    Parameters
+    ----------
+    metamer : po.Metamer
+        The Metamer object to grab data from
+    iteration : int or None, optional
+        Which iteration to display. If None, we show the most recent one.
+        Negative values are also allowed.
+    pixel_value_subsample : float, optional
+        Float between 0 and 1 giving the percentage of image pixels to plot.
+        Done to make visualization clearer.
+
+    Returns
+    -------
+    fig : plt.Figure
+        Figure containing the plot
+    axes_idx : dict
+        dictionary specifying which plot is where, for use with animate()
+
+    Notes
+    -----
+    To successfully `animate_schematic`, pass the same metamer object,
[…]
