### 2023-11-28 3a7b771
synthesis_video.py now runnable from command line
