### 2023-11-28 b8a0455
adds synthesis_video to tests
