### 2023-11-28 b8133f1
save synthesis schematic figure as well
