### 2023-12-05 #150 @sneakers-the-rat
## Editor response to review:

---

## Editor comments
:wave: Hi @NickleDave  and @\yet-to-be-named-reviewer-2 Thank you for volunteering to review
for pyOpenSci! I'm dropping this editor template here now and will update it once we find the second reviewer. Reached out to a few folks now and will continue to do so until we find a second!

- [ ] reviewer 3 (if applicable)

Please get in touch with any questions or concerns! Your review is due: **no deadline yet still finding reviewers**
---

Reviewers: @NickleDave and @\unknown!!!!
Due date: TBD
