### 2023-12-11 4b607e9
adds info about pip minimum version
