### 2023-12-14 #150 @billbrod
Thanks all, looking forward to this process! While I'm looking forward to all kinds of feedback,  I am specifically interested in feedback on the docs, especially the motivation for the included methods and explaining what they can be used for scientifically -- I think that's the weakest (and most important?) part of our documentation right now. Since all the primary authors for the project are from the same lab, we talked a lot offline and have a similar background.
