### 2023-12-14 2f762b8
adds more details to install instructions
more about jupyter, optional deps
