+++ docs/install.rst
@@ -7,16 +7,13 @@ The following instructions will work on Linux or Mac. If you're on Windows, I
 recommend looking into the `Windows Subsystem for Linux
 <https://docs.microsoft.com/en-us/windows/wsl/install-win10).>`_
 
-The easiest way to install ``plenoptic`` is with pip within a new `conda
-<https://docs.conda.io/en/latest/>`_ environment (if you do not have ``conda``
-installed on your machine, I recommend starting with `miniconda
-<https://docs.conda.io/en/latest/miniconda.html>`_)::
+The easiest way to install ``plenoptic`` is from `PyPI <https://pypi.org/project/plenoptic/>`_  (the Python Package Index) using pip within a new `conda <https://docs.conda.io/en/latest/>`_ environment (if you do not have ``conda`` installed on your machine, I recommend starting with `miniconda <https://docs.conda.io/en/latest/miniconda.html>`_)::
 
 $ conda create --name plenoptic pip python=3.9
 $ conda activate plenoptic
 $ pip install plenoptic
 
-You can also install it directly from source::
+You can also install it directly from source to have a local editable copy. This is most useful for developing (for more info, see `our contributing guide <https://github.com/LabForComputationalVision/plenoptic/blob/main/CONTRIBUTING.md>`_) or if you want to use the most cutting-edge version::
 
 $ conda create --name plenoptic pip python=3.9
 $ conda activate plenoptic
@@ -26,7 +23,21 @@ $ cd plenoptic
 $ # install in editable mode with `-e` or, equivalently, `--editable`
 $ pip install -e .
 
+With an editable copy, any changes locally will be automatically reflected in your installation (under the hood, this command uses symlinks).
+
 .. attention:: To install ``plenoptic`` in editable mode, you need ``pip >= 21.3`` (see pip's `changelog <https://pip.pypa.io/en/stable/news/#id286>`_). If you run into `an error <https://github.com/LabForComputationalVision/plenoptic/issues/227>`_ after running the ``pip install -e .`` command, try updating your pip version with ``pip install --upgrade pip``.
+
+.. _optional-deps:
+Optional dependencies
+---------------------
+
+The above instructions will install plenoptic and its core dependencies. You may also wish to install some additional optional dependencies. These dependencies are specified using square brackets during the pip install command and can be installed for either a local, editable install or one directly from PyPI:
+
+* If you would like to run the jupyter notebooks locally: ``pip install plenoptic[nb]`` or ``pip install -e .[nb]``. This includes ``torchvision``, ``jupyter``, and related libraries. See the :ref:`jupyter section <jupyter>` for more details on how to handle jupyter and python virtual environments.
+* If you would like to locally build the documentation: ``pip install -e .[docs]``. This includes ``sphinx`` and related libraries. (This probably only makes sense if you have a local installation.)
+* If you would like to run the tests: ``pip install -e .[dev]``. This includes ``pytest`` and related libraries. (This probably only makes sense if you have a local installation.)
+
+These optional dependencies can be joined with a comma: ``pip install -e .[docs,dev]``
 
 ffmpeg and videos
 -----------------
@@ -45,15 +56,16 @@ To change the backend, run ``matplotlib.rcParams['animation.writer'] = writer``
 before calling any of the animate functions. If you try to set that ``rcParam``
 with a random string, ``matplotlib`` will tell you the available choices.
 
-Jupyter
--------
+.. _jupyter:
+Running notebooks locally
+-------------------------
 
 If you wish to locally run the notebooks, you will need to install ``jupyter``,
-``ipywidgets``, and (for some of the notebooks) ``torchvision`` (you can also
-run them in the cloud using `Binder
-<https://mybinder.org/v2/gh/LabForComputationalVision/plenoptic/1.0.1?filepath=examples>`_).
[…]
