+++ pyproject.toml
@@ -56,6 +56,7 @@ dev = [
 nb = [
     'jupyter',
     'ipywidgets',
+    "torchvision>=0.3",
     'nbclient>=0.5.5',
 ]
 
