### 2023-12-14 bedabdf
updates CONTRIBUTING
more info about how branching and contribution workflow works
