+++ CONTRIBUTING.md
@@ -2,22 +2,19 @@
 
 `plenoptic` is a python library of tools to help researchers better understand
 their models. While the majority of authors are members or alumni of NYU's [Lab
-for Computational Vision](https://www.cns.nyu.edu/~lcv/), we welcome other
-contributors!
+for Computational Vision](https://www.cns.nyu.edu/~lcv/), we are open to
+contributions from anyone!
 
 First, please check out the [Code of Conduct](CODE_OF_CONDUCT.md) and read it
-before going any further. You may also want to check out the [README](README.md)
-for a longer overview of the project and how to get everything installed, the
-tutorials (Jupyter notebooks found in the `examples/` folder) for some examples
-of how to interact with the library, and the
-[documentation](https://plenoptic.readthedocs.io/).
+before going any further. You may also want to check out the [main page of the
+documentation](https://plenoptic.readthedocs.io/en/latest/) for a longer
+overview of the project and how to get everything installed, as well as pointers
+for further reading, depending on your interests.
 
 If you encounter any issues with `plenoptic`, please open an
-[issue](https://github.com/LabForComputationalVision/plenoptic/issues)!
-We have a template for bug reports, following it (so you provide the
-necessary details) will make solving it easier. Right now, we have
-enough on our plate that we're not considering any enhancements or new
-features --- we're focusing on implementing what we plan to do.
+[issue](https://github.com/LabForComputationalVision/plenoptic/issues)! We have
+a template for bug reports, and following it (so you provide the necessary
+details) will make solving your problem much easier.
 
 If you'd like to help improve `plenoptic`, we have a bunch of issues
 we're working through. For people who are not already familiar with
@@ -28,8 +25,11 @@ make progress on one of the existing issues, please give it a try.
 
 In order to submit changes, create a branch or fork of the project, make your
 changes, add documentation and tests, and submit a [Pull
-Request](https://github.com/LabForComputationalVision/plenoptic/pulls). The
-amount and form of documentation to add depends on the size of the submitted
+Request](https://github.com/LabForComputationalVision/plenoptic/pulls). See
+[contributing to the code below](#contributing-to-the-code) for more details on
+this process.
+
+The amount and form of documentation to add depends on the size of the submitted
 changes. For a significant change (a new model or synthesis method), please
 include a new tutorial notebook that walks through how to use them. For
 enhancements of existing methods, you can probably just modify the existing
@@ -55,28 +55,64 @@ We try to keep all our communication on Github, and we use several channels:
 
 ## Contributing to the code
 
+### Contribution workflow
+
 We welcome contributions to `plenoptic`! In order to contribute, please create
-your own branch, make sure the tests pass, and open a Pull Request. If you're a
-member of the [LCV github](https://github.com/LabForComputationalVision/), you
-can create a branch directly in the repository: from within the `plenoptic`
-repository:
+your own branch, make sure the tests pass, and open a Pull Request. We follow
+the [GitHub
+Flow](https://www.gitkraken.com/learn/git/best-practices/git-branch-strategy#github-flow-branch-strategy)
+workflow: no one is allowed to push to the `main` branch, all development
+happens in separate feature branches (each of which, ideally, implements a
+single feature, addresses a single issue, or fixes a single problem), and these
+get merged into `main` once we have determined they're ready. Then, after enough
+changes have accumulated, we put out a new release, adding a new tag which
+increments the version number, and uploading the new release to PyPI (see
+[releases](#releases) for more details).
+
[…]
