+++ CITATION.cff
@@ -14,12 +14,12 @@ authors:
   - family-names: Fiquet
     given-names: Pierre-Étienne
     orcid: "https://orcid.org/0000-0002-8301-2220"
-  - family-names: Parthasarathy
-    given-names: Nikhil
-    orcid: "https://orcid.org/0000-0003-2572-6492"
   - family-names: Herrera-Esposito
     given-names: Daniel
     orcid: "https://orcid.org/0000-0001-8181-9787"
+  - family-names: Parthasarathy
+    given-names: Nikhil
+    orcid: "https://orcid.org/0000-0003-2572-6492"
   - family-names: Simoncelli
     given-names: Eero
     orcid: "https://orcid.org/0000-0002-1206-527X"
