+++ pyproject.toml
@@ -62,7 +62,7 @@ nb = [
 [project.urls]
 "Homepage" = "https://github.com/LabForComputationalVision/plenoptic"
 "Documentation" = "https://plenoptic.readthedocs.io/en/latest/"
-"Download" = "https://zenodo.org/record/3995056"
+"Download" = "https://zenodo.org/doi/10.5281/zenodo.10151130"
 
 [tool.setuptools.packages.find]
 where = ["src"]
