### 2023-12-28 #150 @sneakers-the-rat
This is ur friendly neighborhood editor here to say that I hope the authors and reviewers are taking a much deserved break, and to keep a reasonable and humane timeline I am making the call to reset the review window to start Jan 1 - Lets say reviews in **January 19th, 2024**

Have a lovely rest of the holiday
