+++ src/plenoptic/simulate/canonical_computations/steerable_pyramid_freq.py
@@ -144,16 +144,9 @@ class SteerablePyramidFreq(nn.Module):
         # create low and high masks
         lo0mask = interpolate1d(self.log_rad, self.YIrcos, self.Xrcos)
         hi0mask = interpolate1d(self.log_rad, self.Yrcos, self.Xrcos)
-        self.lo0mask = torch.tensor(lo0mask).unsqueeze(0)
-        self.hi0mask = torch.tensor(hi0mask).unsqueeze(0)
+        self.register_buffer('lo0mask', torch.tensor(lo0mask).unsqueeze(0))
+        self.register_buffer('hi0mask', torch.tensor(hi0mask).unsqueeze(0))
 
-        # pre-generate the angle, hi and lo masks, as well as the
-        # indices used for down-sampling
-        self._anglemasks = []
-        self._anglemasks_recon = []
-        self._himasks = []
-        self._lomasks = []
-        self._loindices = []
 
         # need a mock image to down-sample so that we correctly
         # construct the differently-sized masks
@@ -171,6 +164,9 @@ class SteerablePyramidFreq(nn.Module):
         Xrcos = self.Xrcos.copy()
         angle = self.angle.copy()
         log_rad = self.log_rad.copy()
+        # pre-generate the angle, hi and lo masks, as well as the
+        # indices used for down-sampling.
+        self._loindices = []
         for i in range(self.num_scales):
             Xrcos -= np.log2(2)
             const = ((2 ** (2*self.order)) * (factorial(self.order, exact=True)**2) /
@@ -187,7 +183,7 @@ class SteerablePyramidFreq(nn.Module):
                 Ycosn_recon = Ycosn_forward
 
             himask = interpolate1d(log_rad, self.Yrcos, Xrcos)
-            self._himasks.append(torch.tensor(himask).unsqueeze(0))
+            self.register_buffer(f'_himasks_scale_{i}', torch.tensor(himask).unsqueeze(0))
 
             anglemasks = []
             anglemasks_recon = []
@@ -199,11 +195,11 @@ class SteerablePyramidFreq(nn.Module):
                 anglemasks.append(torch.tensor(anglemask).unsqueeze(0))
                 anglemasks_recon.append(torch.tensor(anglemask_recon).unsqueeze(0))
 
-            self._anglemasks.append(anglemasks)
-            self._anglemasks_recon.append(anglemasks_recon)
+            self.register_buffer(f'_anglemasks_scale_{i}', torch.cat(anglemasks))
+            self.register_buffer(f'_anglemasks_recon_scale_{i}', torch.cat(anglemasks_recon))
             if not self.downsample:
                 lomask = interpolate1d(log_rad, self.YIrcos, Xrcos)
-                self._lomasks.append(torch.tensor(lomask).unsqueeze(0))
+                self.register_buffer(f'_lomasks_scale_{i}', torch.tensor(lomask).unsqueeze(0))
                 self._loindices.append([np.array([0, 0]), dims])
                 lodft = lodft * lomask
 
@@ -222,62 +218,14 @@ class SteerablePyramidFreq(nn.Module):
                 angle = angle[lostart[0]:loend[0], lostart[1]:loend[1]]
 
                 lomask = interpolate1d(log_rad, self.YIrcos, Xrcos)
-                self._lomasks.append(torch.tensor(lomask).unsqueeze(0))
+                self.register_buffer(f'_lomasks_scale_{i}', torch.tensor(lomask).unsqueeze(0))
                 # subsampling
                 lodft = lodft[lostart[0]:loend[0], lostart[1]:loend[1]]
                 # convolution in spatial domain
                 lodft = lodft * lomask
 
         # reasonable default dtype
-        self = self.to(torch.float32)
-
[…]
