### 2024-01-03 8ee8e84
updates OS note
