### 2024-01-03 f9f230a
Apply Edoardo's suggestions from code review
Co-authored-by: Edoardo Balzani <edoardo.balzani87@gmail.com>
