+++ src/plenoptic/simulate/models/portilla_simoncelli.py
@@ -146,8 +146,7 @@ class PortillaSimoncelli(nn.Module):
         scales_without_coarsest = np.arange(self.n_scales-1)
         # the statistics computed on the reconstructed bandpass images have an
         # extra scale corresponding to the lowpass residual
-        scales_with_lowpass = np.concatenate([scales, ["residual_lowpass"]],
-                                             dtype=object)
+        scales_with_lowpass = np.array(scales.tolist() + ["residual_lowpass"], dtype=object)
 
         # now we go through each statistic in order and create a dummy array
         # full of 1s with the same shape as the actual statistic (excluding the
@@ -242,7 +241,7 @@ class PortillaSimoncelli(nn.Module):
             if k in ["auto_correlation_magnitude", "auto_correlation_reconstructed"]:
                 # Symmetry M_{i,j} = M_{n-i+1, n-j+1}
                 # Start with all False, then place True in necessary stats.
-                mask = torch.zeros(v.shape, dtype=bool)
+                mask = torch.zeros(v.shape, dtype=torch.bool)
                 mask[tril_inds[0], tril_inds[1]] = True
                 # if spatial_corr_width is even, then the first row is not
                 # redundant with anything either
@@ -252,11 +251,11 @@ class PortillaSimoncelli(nn.Module):
             elif k == 'cross_orientation_correlation_magnitude':
                 # Symmetry M_{i,j} = M_{j,i}.
                 # Start with all True, then place False in redundant stats.
-                mask = torch.ones(v.shape, dtype=bool)
+                mask = torch.ones(v.shape, dtype=torch.bool)
                 mask[triu_inds[0], triu_inds[1]] = False
             else:
                 # all of the other stats have no redundancies
-                mask = torch.ones(v.shape, dtype=bool)
+                mask = torch.ones(v.shape, dtype=torch.bool)
             mask_dict[k] = mask
         return mask_dict
 
@@ -336,7 +335,7 @@ class PortillaSimoncelli(nn.Module):
         # reconstructed lowpass image. std_recon, skew_recon, and
         # kurtosis_recon will all end up as tensors of shape (batch, channel,
         # n_scales+1)
-        std_recon = var_recon.pow(0.5)
+        std_recon = var_recon.sqrt()
         skew_recon, kurtosis_recon = self._compute_skew_kurtosis_recon(reconstructed_images, var_recon, pixel_stats[..., 1])
 
         # Compute the cross-orientation correlations between the magnitude
@@ -348,7 +347,7 @@ class PortillaSimoncelli(nn.Module):
         # (it's an intermediary of the computation of the cross-orientation
         # correlations), of shape (batch, channel, n_orientations, n_scales).
         # We take the square root to get the standard deviation.
-        mags_std = mags_var.pow(0.5)
+        mags_std = mags_var.sqrt()
 
         # If we have more than one scale, compute the cross-scale correlations
         if self.n_scales != 1:
@@ -401,7 +400,7 @@ class PortillaSimoncelli(nn.Module):
     def remove_scales(
             self, representation_vector: Tensor, scales_to_keep: List[SCALES_TYPE]
     ) -> Tensor:
-        """Remove statistics not associated with scales
+        """Remove statistics not associated with scales.
 
         For a given representation_vector and a list of scales_to_keep, this
         attribute removes all statistics *not* associated with those scales.
@@ -419,7 +418,7 @@ class PortillaSimoncelli(nn.Module):
             representation corresponding to those scales.
 
         Returns
-        ------
+        -------
         limited_representation_vector :
             Representation vector with some statistics removed.
 
[…]
