### 2024-01-03 be7909d
add update_plot test
