+++ tests/test_models.py
@@ -713,7 +713,20 @@ class TestPortillaSimoncelli(object):
             n_orientations=n_orientations,
             spatial_corr_width=spatial_corr_width,
             ).to(DEVICE)
-        model.plot_representation(model(einstein_img.repeat((*batch_channel, 1, 1))))
+        model.plot_representation(model(einstein_img.repeat((*batch_channel, 1, 1))),
+                                  title="Representation")
+
+    def test_update_plot(self, einstein_img):
+        model = po.simul.PortillaSimoncelli(
+            einstein_img.shape[-2:],
+            ).to(DEVICE)
+        _, axes = model.plot_representation(model(einstein_img))
+        orig_y = axes[0].containers[0].markerline.get_ydata()
+        img = po.load_images(op.join(DATA_DIR, "256", "nuts.pgm"))
+        artists = model.update_plot(axes, model(img))
+        updated_y = artists[0].get_ydata()
+        if np.equal(orig_y, updated_y).all():
+            raise ValueError("Update plot didn't run successfully!")
 
     @pytest.mark.parametrize("batch_channel", [(1, 1), (1, 3), (2, 1), (2, 3)])
     @pytest.mark.parametrize("n_scales", [1, 2, 3, 4])
