### 2024-01-03 02a72ea
adds test for convert to dict errors
