### 2024-01-03 9327bcd
test even spatial_corr_width
need to update regression tests for this
