+++ tests/test_models.py
@@ -460,7 +460,7 @@ def remove_redundant_and_normalize(matlab_rep: OrderedDict,
 class TestPortillaSimoncelli(object):
     @pytest.mark.parametrize("n_scales", [1, 2, 3, 4])
     @pytest.mark.parametrize("n_orientations", [2, 3, 4])
-    @pytest.mark.parametrize("spatial_corr_width", [3, 5, 7, 9])
+    @pytest.mark.parametrize("spatial_corr_width", range(3, 10))
     def test_portilla_simoncelli(
         self,
         n_scales,
@@ -553,7 +553,7 @@ class TestPortillaSimoncelli(object):
 
     @pytest.mark.parametrize("n_scales", [1, 2, 3, 4])
     @pytest.mark.parametrize("n_orientations", [2, 3, 4])
-    @pytest.mark.parametrize("spatial_corr_width", [3, 5, 7, 9])
+    @pytest.mark.parametrize("spatial_corr_width", range(3, 10))
     def test_ps_convert(self, n_scales, n_orientations, spatial_corr_width,
                         einstein_img):
         ps = po.simul.PortillaSimoncelli(
@@ -688,7 +688,7 @@ class TestPortillaSimoncelli(object):
     @pytest.mark.parametrize("batch_channel", [(1, 3), (2, 1), (2, 3)])
     @pytest.mark.parametrize("n_scales", [1, 2, 3, 4])
     @pytest.mark.parametrize("n_orientations", [2, 3, 4])
-    @pytest.mark.parametrize("spatial_corr_width", [3, 5, 7, 9])
+    @pytest.mark.parametrize("spatial_corr_width", range(3, 10))
     def test_multibatchchannel(self, batch_channel, n_scales, n_orientations,
                                spatial_corr_width, einstein_img):
         model = po.simul.PortillaSimoncelli(
@@ -704,7 +704,7 @@ class TestPortillaSimoncelli(object):
     @pytest.mark.parametrize("batch_channel", [(1, 1), (1, 3), (2, 1), (2, 3)])
     @pytest.mark.parametrize("n_scales", [1, 2, 3, 4])
     @pytest.mark.parametrize("n_orientations", [2, 3, 4])
-    @pytest.mark.parametrize("spatial_corr_width", [3, 5, 7, 9])
+    @pytest.mark.parametrize("spatial_corr_width", range(3, 10))
     def test_plot_representation(self, batch_channel, n_scales, n_orientations,
                                  spatial_corr_width, einstein_img):
         model = po.simul.PortillaSimoncelli(
@@ -731,7 +731,7 @@ class TestPortillaSimoncelli(object):
     @pytest.mark.parametrize("batch_channel", [(1, 1), (1, 3), (2, 1), (2, 3)])
     @pytest.mark.parametrize("n_scales", [1, 2, 3, 4])
     @pytest.mark.parametrize("n_orientations", [2, 3, 4])
-    @pytest.mark.parametrize("spatial_corr_width", [3, 5, 7, 9])
+    @pytest.mark.parametrize("spatial_corr_width", range(3, 10))
     def test_plot_representation_dim_assumption(self, batch_channel, n_scales,
                                                 n_orientations, spatial_corr_width,
                                                 einstein_img):
@@ -758,7 +758,7 @@ class TestPortillaSimoncelli(object):
 
     @pytest.mark.parametrize("n_scales", [1, 2, 3, 4])
     @pytest.mark.parametrize("n_orientations", [2, 3, 4])
-    @pytest.mark.parametrize("spatial_corr_width", [3, 5, 7, 9])
+    @pytest.mark.parametrize("spatial_corr_width", range(3, 10))
[…]
