### 2024-01-03 493eb40
updates test_ps_torch_output
create new file for this
