+++ tests/test_metric.py
@@ -17,6 +17,7 @@ OSF_URL = {'plenoptic-test-files.tar.gz': 'q9kn8', 'ssim_images.tar.gz': 'j65tw'
            'ssim_analysis.mat': 'ndtc7', 'msssim_images.tar.gz': '5fuba', 'MAD_results.tar.gz': 'jwcsr',
            'portilla_simoncelli_matlab_test_vectors.tar.gz': 'qtn5y',
            'portilla_simoncelli_test_vectors.tar.gz': '8r2gq',
+           'portilla_simoncelli_test_vectors_refactor.tar.gz': 'ca7qt',
            'portilla_simoncelli_images.tar.gz':'eqr3t',
            'portilla_simoncelli_synthesize.npz': 'a7p9r',
            'portilla_simoncelli_synthesize_torch_v1.12.0.npz': 'gbv8e',
@@ -44,6 +45,7 @@ def osf_download(filename):
                 'portilla_simoncelli_images.tar.gz',
                 'portilla_simoncelli_matlab_test_vectors.tar.gz',
                 'portilla_simoncelli_test_vectors.tar.gz',
+                'portilla_simoncelli_test_vectors_refactor.tar.gz',
                 'portilla_simoncelli_synthesize.npz',
                 'portilla_simoncelli_synthesize_torch_v1.12.0.npz',
                 'portilla_simoncelli_synthesize_torch_v1.12.0_ps-refactor.npz',
