### 2024-01-03 ce13d8b
updates test_ps_scales to include even spatial corr width
