### 2024-01-03 95ca562
makes pyramid attribute private
