### 2024-01-03 514acf3
replace B,C,S / B,C,H,W style notation
