### 2024-01-03 f7a575d
change type annotations for Figure/Axes
