### 2024-01-04 c4ee4bd
adds explanation for einops.rearrange
