+++ src/plenoptic/simulate/canonical_computations/steerable_pyramid_freq.py
@@ -10,11 +10,12 @@ from einops import rearrange
 from scipy.special import factorial
 from torch import Tensor
 from typing_extensions import Literal
+from numpy.typing import NDArray
 
 from ...tools.signal import interpolate1d, raised_cosine, steer
 
 complex_types = [torch.cdouble, torch.cfloat]
-SCALES_TYPE = List[Union[int, Literal["residual_lowpass", "residual_highpass"]]]
+SCALES_TYPE = Union[int, Literal["residual_lowpass", "residual_highpass"]]
 KEYS_TYPE = Union[Tuple[int, int], Literal["residual_lowpass", "residual_highpass"]]
 
 
@@ -311,7 +312,7 @@ class SteerablePyramidFreq(nn.Module):
     def forward(
         self,
         x: Tensor,
-        scales: Optional[SCALES_TYPE] = None,
+        scales: Optional[List[SCALES_TYPE]] = None,
     ) -> OrderedDict:
         r"""Generate the steerable pyramid coefficients for an image
 
@@ -442,7 +443,7 @@ class SteerablePyramidFreq(nn.Module):
     @staticmethod
     def convert_pyr_to_tensor(
         pyr_coeffs: OrderedDict, split_complex: bool = False
-    ) -> Tuple[Tensor, Tuple[int, bool, SCALES_TYPE]]:
+    ) -> Tuple[Tensor, Tuple[int, bool, List[KEYS_TYPE]]]:
         r"""Convert coefficient dictionary to a tensor.
 
         The output tensor has shape (batch, channel, height, width) and is
@@ -489,7 +490,7 @@ class SteerablePyramidFreq(nn.Module):
 
         """
 
-        pyr_keys = tuple(pyr_coeffs.keys())
+        pyr_keys = list(pyr_coeffs.keys())
         test_band = pyr_coeffs[pyr_keys[0]]
         num_channels = test_band.size(1)
         coeff_list = []
@@ -532,7 +533,7 @@ class SteerablePyramidFreq(nn.Module):
         pyr_tensor: Tensor,
         num_channels: int,
         split_complex: bool,
-        pyr_keys: SCALES_TYPE,
+        pyr_keys: List[KEYS_TYPE],
     ) -> OrderedDict:
         r"""Convert pyramid coefficient tensor to dictionary format.
 
@@ -600,7 +601,7 @@ class SteerablePyramidFreq(nn.Module):
         return pyr_coeffs
 
     def _recon_levels_check(
-        self, levels: Union[SCALES_TYPE, Literal["all"], List[SCALES_TYPE]]
+        self, levels: Union[Literal["all"], List[SCALES_TYPE]]
     ) -> List[SCALES_TYPE]:
         r"""Check whether levels arg is valid for reconstruction and return valid version
 
@@ -616,8 +617,7 @@ class SteerablePyramidFreq(nn.Module):
             If `list` should contain some subset of integers from `0` to
             `self.num_scales-1` (inclusive) and `'residual_highpass'` and
             `'residual_lowpass'` (if appropriate for the pyramid). If `'all'`,
-            returned value will contain all valid levels. Otherwise, must be
-            one of the valid levels.
+            returned value will contain all valid levels.
 
         Returns
         -------
@@ -625,18 +625,19 @@ class SteerablePyramidFreq(nn.Module):
             List containing the valid levels for reconstruction.
 
         """
-        if isinstance(levels, str) and levels == "all":
+        if isinstance(levels, str):
+            if levels != "all":
+                raise TypeError(f"levels must be a list of levels or the string 'all' but got {levels}")
             levels = (
                 ["residual_highpass"]
                 + list(range(self.num_scales))
                 + ["residual_lowpass"]
             )
         else:
[…]
