### 2024-01-04 78144d2
change threshold to be based on dtype.resolution
