### 2024-01-04 28b3874
adds module level docstrings
