### 2024-01-05 ff6ca27
updates contributing / install in response to Eric's comments
