+++ docs/install.rst
@@ -58,42 +58,49 @@ with a random string, ``matplotlib`` will tell you the available choices.
 Running notebooks locally
 -------------------------
 
+.. tip:: You can run the notebooks in the cloud using `Binder <https://mybinder.org/v2/gh/LabForComputationalVision/plenoptic/1.0.1?filepath=examples>`_, no installation required!
+
+Installing jupyter and setting up the kernel
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
 If you wish to locally run the notebooks, you will need to install ``jupyter``,
 ``ipywidgets``, and (for some of the notebooks) ``torchvision`` .
-There are three possible of getting a local jupyter install working with this
-package, depending on how you wish to handle virtual environments:
+There are three possible ways of getting a local jupyter install working with this
+package, depending on how you wish to handle virtual environments.
 
-.. tip:: You can run the notebooks in the cloud using `Binder <https://mybinder.org/v2/gh/LabForComputationalVision/plenoptic/1.0.1?filepath=examples>`_, no installation required! (This link is also found in the binder badge on the `README <https://github.com/LabForComputationalVision/plenoptic/>`_ or :ref:`main docs page <index>`.)
+.. hint:: If ``plenoptic`` is the only environment that you want to run notebooks from and/or you are unfamiliar with virtual environments, go with option 1 below.
 
-1. Install jupyter in the same environment as ``plenoptic``. If you followed the
-   instructions above to create a ``conda`` environment named ``plenoptic``, do
-   the following::
+1. Install jupyter in the same environment as ``plenoptic``. This is the easiest
+   but, if you have multiple virtual environments and want to use Jupyter
+   notebooks in each of them, it will take up a lot of space. If you followed
+   the instructions above to create a ``conda`` environment named ``plenoptic``,
+   do the following::
 
    $ conda activate plenoptic
    $ conda install -c conda-forge jupyterlab ipywidgets torchvision
 
-   This is the easiest but, if you have multiple virtual environments and want
-   to use Jupyter notebooks in each of them, it will take up a lot of space.
+   With this setup, when you have another virtual environment that you wish to run jupyter notebooks from, you must reinstall jupyuter into that separate virtual environment, which is wasteful.
 
-2. Use `nb_conda_kernels
+2. Install jupyter in your ``base`` environment and use `nb_conda_kernels
    <https://github.com/Anaconda-Platform/nb_conda_kernels>`_ to automatically
-   manage kernels in your conda environments. Again, if you followed the
-   instructions to create a ``conda`` environment named ``plenoptic``::
+   manage kernels in all your conda environments. This is a bit more
+   complicated, but means you only have one installation of jupyter lab on your
+   machine. Again, if you followed the instructions to create a ``conda``
+   environment named ``plenoptic``::
 
    $ # activate your 'base' environment, the default one created by conda/miniconda
-   $ conda activate
+   $ conda activate base
    $ # install jupyter lab and nb_conda_kernels in your base environment
    $ conda install -c conda-forge jupyterlab ipywidgets
    $ conda install nb_conda_kernels
    $ # install ipykernel and torchvision in the plenoptic environment
    $ conda install -n plenoptic ipykernel torchvision
 
-   This is a bit more complicated, but means you only have one installation of
-   jupyter lab on your machine.
+   With this setup, you have a single jupyter install that can run kernels from any of your conda environments. All you have to do is install ``ipykernel`` (and restart jupyter) and you should see the new kernel!
 
    .. attention:: This method only works with conda environments. If you are using another method to manage your python virtual environments, you'll have to use one of the other methods.
 
[…]
