+++ CONTRIBUTING.md
@@ -1,9 +1,7 @@
 # CONTRIBUTING
 
 `plenoptic` is a python library of tools to help researchers better understand
-their models. While the majority of authors are members or alumni of NYU's [Lab
-for Computational Vision](https://www.cns.nyu.edu/~lcv/), we are open to
-contributions from anyone!
+their models. We welcome and encourage contributions from everyone!
 
 First, please check out the [Code of Conduct](CODE_OF_CONDUCT.md) and read it
 before going any further. You may also want to check out the [main page of the
@@ -11,10 +9,23 @@ documentation](https://plenoptic.readthedocs.io/en/latest/) for a longer
 overview of the project and how to get everything installed, as well as pointers
 for further reading, depending on your interests.
 
-If you encounter any issues with `plenoptic`, please open an
+If you encounter any issues with `plenoptic`, first search the existing
+[issues](https://github.com/LabForComputationalVision/plenoptic/issues) and
+[discussions](https://github.com/LabForComputationalVision/plenoptic/discussions)
+to see if there's already information to help you. If not, please open a new
 [issue](https://github.com/LabForComputationalVision/plenoptic/issues)! We have
 a template for bug reports, and following it (so you provide the necessary
 details) will make solving your problem much easier.
+
+If you'd like to help improve `plenoptic`, there are many ways you can
+contribute, from improving documentation to writing code. For those not already
+familiar with the project, it can be very helpful for us if you go through the
+tutorials, README, and documentation and let us know if anything is unclear,
+what needs more detail (or clearer writing), etc. For those that want to
+contribute code, we also have many
+[issues](https://github.com/LabForComputationalVision/plenoptic/issues) that we
+are working through. If you would like to work on one of those, please give it a
+try!
 
 If you'd like to help improve `plenoptic`, we have a bunch of issues
 we're working through. For people who are not already familiar with
@@ -28,17 +39,6 @@ changes, add documentation and tests, and submit a [Pull
 Request](https://github.com/LabForComputationalVision/plenoptic/pulls). See
 [contributing to the code below](#contributing-to-the-code) for more details on
 this process.
-
-The amount and form of documentation to add depends on the size of the submitted
-changes. For a significant change (a new model or synthesis method), please
-include a new tutorial notebook that walks through how to use them. For
-enhancements of existing methods, you can probably just modify the existing
-tutorials and add documentation. If unsure, ask! For docstrings, we follow
-[numpydoc](https://numpydoc.readthedocs.io/en/latest/format.html) style. See
-later in this file for more details on how to run the tests and build the
-documentation (if you create a branch on the main repo, Github Actions will run
-tests automatically whenever you push, so you don't need to worry about running
-them locally).
 
 We try to keep all our communication on Github, and we use several channels:
 
@@ -69,7 +69,7 @@ changes have accumulated, we put out a new release, adding a new tag which
 increments the version number, and uploading the new release to PyPI (see
 [releases](#releases) for more details).
 
-In addition to the information that follows, [Github](https://docs.github.com/en/get-started/quickstart/github-flow) (unsurprisingly) has good information on this workflow, as does the [Caiman package](https://github.com/flatironinstitute/CaImAn/blob/main/CONTRIBUTING.md) (though note that they use **git** flow, which involves a separate develop branch in addition to main).
+In addition to the information that follows, [Github](https://docs.github.com/en/get-started/quickstart/github-flow) (unsurprisingly) has good information on this workflow, as does the [Caiman package](https://github.com/flatironinstitute/CaImAn/blob/main/CONTRIBUTING.md) (though note that the Caiman uses **git** flow, which involves a separate develop branch in addition to main).
 
[…]
