+++ pyproject.toml
@@ -43,6 +43,7 @@ docs = [
      # https://nbsphinx.readthedocs.io/en/0.6.0/installation.html#Pygments-Lexer-for-Syntax-Highlighting
      'ipython',
      'sphinx-copybutton',
+     'sphinxemoji',
 ]
 
 dev = [
