### 2024-01-05 9f323f3
rename vector -> tensor, update notebook
adds longer discussion in notebook about what metamers mean in the PS
texture context and why they're useful.

also renames vector -> tensor, because that's clearer
