+++ src/plenoptic/simulate/models/portilla_simoncelli.py
@@ -115,15 +115,15 @@ class PortillaSimoncelli(nn.Module):
         # Dictionary defining necessary statistics, that is, those that are not
         # redundant
         self._necessary_stats_dict = self._create_necessary_stats_dict(scales_shape_dict)
-        # turn this into vector we can use in forward pass. first into a
+        # turn this into tensor we can use in forward pass. first into a
         # boolean mask...
         self._necessary_stats_mask = einops.pack(list(self._necessary_stats_dict.values()), '*')[0]
         # then into a tensor of indices
         self._necessary_stats_mask = torch.where(self._necessary_stats_mask)[0]
 
-        # This vector is composed of the following values: 'pixel_statistics',
+        # This array is composed of the following values: 'pixel_statistics',
         # 'residual_lowpass', 'residual_highpass' and integer values from 0 to
-        # self.n_scales-1. It is the same size as the representation vector
+        # self.n_scales-1. It is the same size as the representation tensor
         # returned by this object's forward method. It must be a numpy array so
         # we can have a mixture of ints and strs (and so we can use np.in1d
         # later)
@@ -286,12 +286,12 @@ class PortillaSimoncelli(nn.Module):
         scales :
             Which scales to include in the returned representation. If None, we
             include all scales. Otherwise, can contain subset of values present
-            in this model's ``scales`` attribute, and the returned vector will
+            in this model's ``scales`` attribute, and the returned tensor will
             then contain the subset corresponding to those scales.
 
         Returns
         -------
-        representation_vector:
+        representation_tensor:
             3d tensor of shape (batch, channel, stats) containing the measured
             texture statistics.
 
@@ -398,7 +398,7 @@ class PortillaSimoncelli(nn.Module):
             all_stats += [cross_scale_corr_mags, cross_scale_corr_real]
         all_stats += [var_highpass_residual]
         # And then pack them into a 3d tensor
-        representation_vector, pack_info = einops.pack(all_stats, 'b c *')
+        representation_tensor, pack_info = einops.pack(all_stats, 'b c *')
 
         # the only time when this is None is during testing, when we make sure
         # that our assumptions are all valid.
@@ -408,38 +408,38 @@ class PortillaSimoncelli(nn.Module):
             self._pack_info = pack_info
         else:
             # Throw away all redundant statistics
-            representation_vector = representation_vector.index_select(-1, self._necessary_stats_mask)
+            representation_tensor = representation_tensor.index_select(-1, self._necessary_stats_mask)
 
         # Return the subset of stats corresponding to the specified scale.
         if scales is not None:
-            representation_vector = self.remove_scales(representation_vector, scales)
+            representation_tensor = self.remove_scales(representation_tensor, scales)
 
-        return representation_vector
+        return representation_tensor
 
     def remove_scales(
-            self, representation_vector: Tensor, scales_to_keep: List[SCALES_TYPE]
+            self, representation_tensor: Tensor, scales_to_keep: List[SCALES_TYPE]
     ) -> Tensor:
         """Remove statistics not associated with scales.
 
-        For a given representation_vector and a list of scales_to_keep, this
+        For a given representation_tensor and a list of scales_to_keep, this
         attribute removes all statistics *not* associated with those scales.
 
         Note that calling this method will always remove statistics.
 
         Parameters
         ----------
-        representation_vector:
+        representation_tensor:
             3d tensor containing the measured representation statistics.
         scales_to_keep:
             Which scales to include in the returned representation. Can contain
             subset of values present in this model's ``scales`` attribute, and
-            the returned vector will then contain the subset of the full
[…]
