### 2024-01-09 #150 @DanielaPamplona
Hi,

Thank you for checking, that is kind of you!

I do have a question...

What is a "vignette" ? I believe I never saw/use it... Or it is implicit in the code...?

Cheers,

Daniela Pamplona

----------------------------------------------------------------------------------------------
Daniela Pamplona
Researcher and Lecturer at Université Paris-Est Créteil Val de Marne
https://danielapamplona.github.io/

Sent: Monday, January 08, 2024 at 7:11 PM
From: "Jonny Saunders" ***@***.***>
To: "pyOpenSci/software-submission" ***@***.***>
Cc: "Daniela Pamplona" ***@***.***>, "Mention" ***@***.***>
Subject: Re: [pyOpenSci/software-submission] Plenoptic (Issue #150)

Hello dear reviewers @NickleDave @DanielaPamplona - checking in and seeing how things are going, if any guidance is needed or if i can be of any help getting reviews rolling!
