### 2024-01-09 63c7ecb
change tests to avoid duplicating code
