### 2024-01-09 7f724b0
use ModuleList in steerpyr
