### 2024-01-09 25b48ca
use single function for cpu/cuda test
adds new ALL_MODELS to test
