+++ tests/test_models.py
@@ -17,6 +17,23 @@ import os
 os.environ['CUBLAS_WORKSPACE_CONFIG'] = ':4096:8'
 
 
+ALL_MODELS = [
+    "LPyr",
+    "SPyr",
+    "frontend.LinearNonlinear",
+    "frontend.LuminanceGainControl",
+    "frontend.LuminanceContrastGainControl",
+    "frontend.OnOff",
+    "naive.Identity",
+    "naive.Linear",
+    "naive.Gaussian",
+    "naive.CenterSurround",
+    "PortillaSimoncelli",
+    "Identity",
+    "NLP",
+]
+
+
 @pytest.fixture()
 def image_input():
     return torch.rand(1, 1, 100, 100)
@@ -72,32 +89,39 @@ def portilla_simoncelli_scales():
     return osf_download('portilla_simoncelli_scales.npz')
 
 
-def template_test_cuda(model, img):
+@pytest.mark.parametrize("model", ALL_MODELS, indirect=True)
+@pytest.mark.skipif(DEVICE.type == 'cpu', reason="Can only test on cuda")
+def test_cuda(model, einstein_img):
     model.cuda()
-    model(img)
+    model(einstein_img)
     # make sure it ends on same device it started, since it might be a fixture
     model.to(DEVICE)
 
-def template_test_cpu_and_back(model, img):
+@pytest.mark.parametrize("model", ALL_MODELS, indirect=True)
+@pytest.mark.skipif(DEVICE.type == 'cpu', reason="Can only test on cuda")
+def test_cpu_and_back(model, einstein_img):
     model.cpu()
     model.cuda()
-    model(img)
+    model(einstein_img)
     # make sure it ends on same device it started, since it might be a fixture
     model.to(DEVICE)
 
-def template_test_cuda_and_back(model, img):
+@pytest.mark.parametrize("model", ALL_MODELS, indirect=True)
+@pytest.mark.skipif(DEVICE.type == 'cpu', reason="Can only test on cuda")
+def test_cuda_and_back(model, einstein_img):
     model.cuda()
     model.cpu()
-    model(img.cpu())
+    model(einstein_img.cpu())
     # make sure it ends on same device it started, since it might be a fixture
-    img.to(DEVICE)
+    einstein_img.to(DEVICE)
     model.to(DEVICE)
 
-def template_test_cpu(model, img):
+@pytest.mark.parametrize("model", ALL_MODELS, indirect=True)
+def test_cpu(model, einstein_img):
     model.cpu()
-    model(img.cpu())
+    model(einstein_img.cpu())
     # make sure it ends on same device it started, since it might be a fixture
-    img.to(DEVICE)
+    einstein_img.to(DEVICE)
     model.to(DEVICE)
 
 
@@ -159,25 +183,6 @@ class TestLaplacianPyramid(object):
             # after upsampling up to one row/column. This causes inconsistency on the right and
             # bottom edges, so they are exluded in the comparison.
 
-    @pytest.mark.skipif(DEVICE.type == 'cpu', reason="Can only test on cuda")
-    def test_cuda(self, einstein_img):
-        lpyr = po.simul.LaplacianPyramid()
-        template_test_cuda(lpyr, einstein_img)
-
-    @pytest.mark.skipif(DEVICE.type == 'cpu', reason="Can only test on cuda")
-    def test_cpu_and_back(self, einstein_img):
-        lpyr = po.simul.LaplacianPyramid()
-        template_test_cpu_and_back(lpyr, einstein_img)
-
-    @pytest.mark.skipif(DEVICE.type == 'cpu', reason="Can only test on cuda")
-    def test_cuda_and_back(self, einstein_img):
-        lpyr = po.simul.LaplacianPyramid()
[…]
