### 2024-01-09 2a78716
adds test_validate_model, using ALL_MODELS
