### 2024-01-10 #150 @DanielaPamplona
Got it! Thanks!

----------------------------------------------------------------------------------------------
Daniela Pamplona
Researcher and Lecturer at Université Paris-Est Créteil Val de Marne
https://danielapamplona.github.io/

Sent: Wednesday, January 10, 2024 at 5:04 AM
From: "Jonny Saunders" ***@***.***>
To: "pyOpenSci/software-submission" ***@***.***>
Cc: "Daniela Pamplona" ***@***.***>, "Mention" ***@***.***>
Subject: Re: [pyOpenSci/software-submission] Plenoptic (Issue #150)

@DanielaPamplona

I do have a question... What is a "vignette" ?

Completely fair question. That's a term from R (see Vignettes), so more than reasonable to have not seen it before. They are sort of like longer, narrative code examples, but they don't have a strict definition - if it's easier, you can think of them as synonymous with examples.

If one were to draw the distinction, an example might look like something you'd find in a docstring like:

def myFunction(arg):
    """
    Do a function thing

    Args:
        arg: The thing we give this function

    Example:
        To use this function, give it an argument!

        >>> arg = 1
        >>> print(myFunction(arg))
        5

        Any argument will do!

        >>> arg = 'new argument'
        >>> print(myFunction(arg))
        5

    """
    # TODO: use the argument
    return 5

Where a vignette would be a longer-form demonstration of major functionality of the package, often with real-world use that show multiple things working together.

From a quick look, everything I see in the examples directory could be called a vignette, but I also don't think you need to split hairs too-too finely there :)
