### 2024-01-10 2cf5050
remove unnecessary einstein_img from test
