### 2024-01-12 29d46fd
adds more detailed explanation of scales shape dict
