### 2024-01-17 #150 @DanielaPamplona
Hi,
I have a small question for the authors ( @billbrod ), but this is a bit important. Please answer this as soon as possible. I will submit a detailed review later.
I am trying to reproduce the "quick start" and I do not find the image "einstein.pgm". As a matter of fact, I do not find the folder "data". Is it somewhere on the package or you forgot to submit it? Without it, I cannot continue the review of this section...
Thank you!
Daniela Pamplona
