### 2024-01-17 #150 @NickleDave
Hi @DanielaPamplona I find it here:
https://github.com/LabForComputationalVision/plenoptic/blob/main/data/256/einstein.pgm
(I searched on GitHub)

I think the data folder is this one:
https://github.com/LabForComputationalVision/plenoptic/tree/main/data

So it would be in the root of the project if you clone it with git.

Looks like there's a couple of fixtures in the tests that expect it to live there (GitHub search found this for me too)
https://github.com/LabForComputationalVision/plenoptic/blob/de8f9308d0fcd24050a93b326390eb4444b9aff2/tests/conftest.py#L11
https://github.com/LabForComputationalVision/plenoptic/blob/de8f9308d0fcd24050a93b326390eb4444b9aff2/tests/conftest.py#L33

Does that help?
