### 2024-01-17 #150 @billbrod
Hi @DanielaPamplona sorry about that! @NickleDave is right, the data folder lives in the git repo. I do have a question about that -- it lives there because the development of the tutorials was done by folks who all had cloned the full repo, but users generally don't have that. I don't think the `data/` directory should be included in the installation (the files are only used in the notebooks and tests), so do either of you have advice as to how we should provide the directory to users? Should we just explain it in the [installation instructions](https://plenoptic.readthedocs.io/en/latest/install.html) add a little blurb at the top of each notebook that uses those files?
