### 2024-01-17 #150 @NickleDave
I would suggest adding a `data` or `examples` folder to the project with functionality for accessing it from plenoptic itself. I'll raise an issue on the repo with more detail and link it back here
