### 2024-01-17 LabForComputationalVision/plenoptic#235 @NickleDave
Add data/examples folder to plenoptic with functions for accessing
as discussed in the pyOpenSci review:
https://github.com/pyOpenSci/software-submission/issues/150#issuecomment-1896069593
https://github.com/pyOpenSci/software-submission/issues/150#issuecomment-1896125038

it's typical to want data that a user can access.
Basically there's two ways to do this:
1. directly add files to the package, and then access them with [importlib-resources](https://www.youtube.com/watch?v=ZsGFU2qh73E)
2. host files somewhere else (e.g., OSF) and then download them, e.g. with [pooch](https://github.com/fatiando/pooch)

So for example, for (1) you might have
```
./src/plenoptic/data/einsten.pgm
```

So a user can do
```python
img = plenoptic.data.get('einstein')
# do stuff
```

whereas (2) on the other hand would trigger a call to a function that downloads e.g. a tar.gz of metamers off of a project on OSF -- note this requires some more engineering, e.g. picking a default download location

@billbrod there's some more detail here, would love feedback on this if you have any
https://github.com/NickleDave/python-package-guide/commit/277c5769d215d407c56746d3856d1822a560d605

Also note I would recommend **keeping** the data you have for tests even though it involves some duplication.
That's because you don't want your tests/fixtures to depend on your `plenoptic.data.get` function working.
