### 2024-01-17 LabForComputationalVision/plenoptic#235 @billbrod
Thanks for the help and happy to see there's a write-up being worked on for the python package guide! Happy to give feedback, how should I do so? (do you want to open a PR or should I send you feedback directly?)

I currently use OSF to host some files that I use for testing (e.g., the files to check against during some regression tests), which I download using `requests` -- haven't heard of `pooch` before!

The contents of the `data` folder as it exists on github and as would be helpful for users (i.e., not the files used only for tests) are mainly images, and so are not very large. I'll look into this more, but my inclination then is to use improtlib-resources. It seems like the major downside is total file size and using git for binary files -- I'm already using git for those images (*whomp whomp*) and the total size of the directory is about 2MB, so I don't think that's much of a problem. Does that sound reasonable to you?
