### 2024-01-17 LabForComputationalVision/plenoptic#235 @NickleDave
:100: Totally. For an initial implementation I would literally copy + paste the data dir, add an __init__.py and a get module with an eponymously named function, and from inside it access the files with importlib-resources. 2 MB is not great but it's not gigantic--maybe raise a "trim data dir" issue and add it to the backlog
