### 2024-01-18 #150 @sneakers-the-rat
Thanks y'all - yes this is the kind of thing worth raising an issue about (with a link back to this issue). As reviewers, you are in the position of being an "average user," so if you can't figure out how something works, then some improvement can likely be made in the docs, code, or in this case packaging! Go review team go <3
