### 2024-01-18 LabForComputationalVision/plenoptic#235 @sneakers-the-rat
sounds like u got it, but just as an example here's a very low-overhead way of doing it:
https://github.com/sneakers-the-rat/paper-feeds/blob/1015841855c499798fa30a47dde778512841b87e/src/paper_feeds/const.py#L5

In my (non editorial, personal) opinion, ~2MB is a reasonable cost for making onboarding in a quickstart as seamless as possible, esp. in an archival sense where eg. the image host might go down at some point and having some small example images stored along with the package would be good to test basic functionality. obvs if you want to take those out of the package to host separately that is good too, but just to say don't feel too bad about the 2MB :)
