### 2024-01-19 LabForComputationalVision/plenoptic#236 @NickleDave
DOC: Add repo status badge to README [pyOS review]
Hi @billbrod, I'm finishing raising issues for the first round of the [pyOpenSci review](https://github.com/pyOpenSci/software-submission/issues/150).

We have a checkbox for adding a repostatus badge to the badge section in the README.
https://www.repostatus.org/
In your case I would add the "Active" badge
