### 2024-01-19 plenoptic-org/plenoptic#238 @NickleDave
TST: Test fails locally [pyOpenSci review]
This test failed for me locally:
tests/test_metamers.py::TestMetamers::test_stop_criterion

I need to reboot and re-run so I can give you the full log from the pytest run, will do so after I finish the comment with my entire review :slightly_smiling_face:
