### 2024-01-19 LabForComputationalVision/plenoptic#239 @NickleDave
DOC: Clarify what "model" means in README + conceptual introduction [pyOpenSci review]
Both the README and the conceptual intro section in the docs use the term "model", but this is a very general term.

I would strongly suggest making it clear immediately who the target audience for the package is: principally, researchers building models of vision in the field of cognitive computational neuroscience, with some functionality for models of audition

This should be, like, the second sentence of explaining what plenoptic does, right after saying "it's for synthesizing stimuli using models"
