### 2024-01-19 LabForComputationalVision/plenoptic#240 @NickleDave
DOC: clearly state how to cite in citation section of README and docs, point to citation.cff [pyOpenSci review]
As required by the criteria of the pyOpenSci review, there is a citation section in the README.
But it just links to a list of citations in the docs.

I would suggest the following:
Directly state in this section of the README how the code should be cited. E.g., "If you use the code, please cite both the code by DOI as well as the JOV paper. If you are not using the code directly then cite the paper". I would also take advantage of the citation.cff, by saying immediately after "you can also click on 'cite this repository' on the right side of the page to get a copyable citation". I would then include both the Zenodo badge with the code DOI and a bibtex citation for the paper in this section, just in case someone misses the part about "click on the right side of the page". Finally I would say "for more in-depth citation lists go here (link to citation page)" and I would repeat all this info at the top of that page (without the citation.cff part, obvs)
