### 2024-01-19 LabForComputationalVision/plenoptic#238 @billbrod
When you do so, can you please also include your system info (as requested in our bug report issue template)? (These tests are passing in our CI, so will need to try and track it down)

**System (please complete the following information):**
 - OS: [e.g. Mac (with version), Ubuntu 18.04]
 - Python version [e.g. 3.7]
 - Pytorch version [e.g., 1.4]
 - Plenoptic version [e.g. 0.1]
