### 2024-01-19 #150 @NickleDave
Hi all, please find below my review.

@billbrod I would like to go through some of the docs in more detail, to give feedback on how domain-specific material is presented; I will raise additional issues then. But I want to get this in on time, and raise the initial issues for required items below.
Boxes that are not yet checked will be explained in "required", and I will raise an issue for each item.

##### Readme file  requirements

---

##### Required
In this section I will put to-dos for items in the checklist above that are not checked\
- Documentation
    - [x] **Examples** for all user-facing functions https://github.com/LabForComputationalVision/plenoptic/issues/237
        - I find examples for some methods on main classes, e.g. `plenoptic.synthesize.Metamer.load`
        - but many functions **don't** have examples, and some of these functions end up being core to the package, e.g. `plenoptic.tools.remove_grad`
- README
        - [ ] A [repostatus.org](https://www.repostatus.org/) badge, https://github.com/LabForComputationalVision/plenoptic/issues/236
        - I think it would be worth (1) stating clearly the dependency on torch, here and in the installation page (see below) and (2) linking related libraries that are dependencies, e.g. https://github.com/LabForComputationalVision/pyrtools
- Usability
    - this repeats the to-do from documentation above; I know it can be really tedious to write examples for *literally* every user-facing function but I would suggest that for frequently used functions such as `plenoptic.tools.remove_grad` there should be an example that passes when doctests are run

- Functionality
        - [x] this test failed for me locally: tests/test_metamers.py::TestMetamers::test_stop_criterion https://github.com/LabForComputationalVision/plenoptic/issues/238
      - currently there's no code formatting or linting (e.g. flake8), AFAICT -- left a comment on an existing issue re: formatters here https://github.com/LabForComputationalVision/plenoptic/issues/51#issuecomment-1900731361

##### suggested
In this section I will put *suggested* changes. Some of these start to get very nit-picky so I won't raise issues for all of them

- documentation
    - [x] be clear about what "model" means in the context of plenoptic: these are models of visual perception, although there are some models of audio perception as well. This should be the second sentence of explaining what plenoptic does, right after saying "it's for synthesizing stimuli using models" https://github.com/LabForComputationalVision/plenoptic/issues/239

- README
    - Badge section
        - In the badge section, the link to Zenodo is just text that says DOI; why not use the Zenodo badge?
            - [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10151131.svg)](https://doi.org/10.5281/zenodo.10151131)
    - citation
        - There is a citation section but it just links to a list of citations in the docs. I would suggest the following
            - Directly state in this section of the README how the code should be cited. E.g., "If you use the code, please cite both the code by DOI as well as the JOV paper. If you are not using the code directly then cite the paper". I would also take advantage of the citation.cff, immediately after say "you can also click on 'cite this repository' on the right side of the page to get a copyable citation". I would then include both the Zenodo badge with the code DOI and a bibtex citation for the paper in this section, just in case someone misses the part about "click on the right page". Finally I would say "for more in-depth citation lists go here (link to citation page)" https://github.com/LabForComputationalVision/plenoptic/issues/240

- API
    - In general, accessing core classes and functions is verbose. I have been guilty of this too and I know this would in a sense be a breaking change, but I would strongly suggest importing core classes and functions at the top-level if possible. I want to be able to just type `po.Metamer` and `po.remove_grad`. It could be fine to leave them where they are but just import them
    - Remember [file structure is an impelmentation detail and flat is better than nested](https://benhoyt.com/writings/python-api-design/#module-and-package-structure)
    - Reasonable people can disagree about this but [I prefer to avoid namespaces with names like `tools` or `util`](http://blog.nicholdav.info/four-tips-structuring-research-python/#finally-dont-name-a-module-packageutils-or-packageutilities), because [everything can go there](https://breadcrumbscollector.tech/stop-naming-your-python-modules-utils/), and as it stands now, you are just forcing users to type `tools` a lot. I would suggest moving modules in tools up a level and import the crucial functions directly into the `plenopitc` namespace from those modules. E.g. `plenoptic.remove_grad`. Or maybe `plenoptic.grad.remove` if you must have a module for it (seems weird that this function lives in `validate`).
- it's also confusing and non-standard to internally import a sub-package at the root package level and change its name: e.g. `synthesize` -> `synth` and `simulate` -> `simul`. I kept thinking I missed an import statement with aliases, and I wasn't sure if there was something else going on with the code organization that I didn't know about. I would strongly suggest removing these internal aliases. If they are just too long, rename the module itself.

- installation
    - [x] I would not suggest installing with pip into a conda environment
        - [ ] because conda recommends only installing with pip *last* after everything else
        - [ ] would it be possible to make a conda-forge package?
            - existing issue: https://github.com/LabForComputationalVision/plenoptic/issues/104
            - my comment: https://github.com/LabForComputationalVision/plenoptic/issues/104#issuecomment-1898993294

- documentation
    - consider switching to newer theme like pydata-sphinx?
    - As stated above, I would strongly suggest making it as obvious as possible that most functionality requires torch
        - in install instructions
    - Some thoughts on organization, as per the diataxis framework: https://github.com/LabForComputationalVision/plenoptic/issues/21#issuecomment-1017933312, https://github.com/LabForComputationalVision/plenoptic/issues/21#issuecomment-1900755529
        - put install + quickstart in a "getting started"
        - put model requirements and citation guide in a "reference"
        - most everything else should go in a "user guide"
        - except API should have its own section
        - reproducibility should go in reference
    - API docs
        - formatting looks off here in some places -- I know that sphinx-autodoc is not the easiest thing to work with but some clean-up here would help it look more approachable
        - I added a typo on this issue: https://github.com/LabForComputationalVision/plenoptic/issues/102#issuecomment-1900764245, will add more there if I find them
- functionality / presentation
I tested by running notebooks, mostly.
I kept notes for the key notebooks as I read through them. This starts to get quite nitpicky, please know I am trying to help with presentation of material. Going to add here for now but I could move this to individual issues.
    - quickstart
        - the function `load_images` in the notebook did not work for me at first
        - I read the function but it uses `op` -- I would suggest making the code as readable as possible by not abbreviating and used absolute names instead, esp for standard library functions; code is read much more than it's written
        - The problem was that the notebook expects to be run from inside notebook and so it has a hardcoded path
            - a fix for this would be to run notebooks from root
            - and/or to include images in the package as discussed in (link issue)
        - I would also suggest re-writing this with pathlib; `pathlib.Path().exists()`
        - and raising an explicit error like "file does not exist: {}"
- packaging guidelines
    - pyproject.toml
        - I would change trove classifiers
            - the package is not in beta annymore if you are using in the lab
            - you can also specify specific pythons: 3.7, etc
        - You probably want to raise the minimum python requires and adhere to [SPEC0](https://scientific-python.org/specs/spec-0000/) since you are depending on core scientific python libraries
        - right now core packages are at 3.10-3.12, but if you need numba then you might need to raise it to 3.9 and test on 3.9-3.11 since [numba is not out for 3.12 just yet](https://github.com/numba/numba/issues/9197)
#### Reading notes
##### conceptual introduction
  - ""> Synthesis is a framework for exploring models" -- what kind of models? (Answer: cognitive/neural!)

-  index
  - "descarded" -> "discarded"

##### quickstart
- "Models can be really simple, as this demonstrates."
    - I think you want to say something here like "Models can be really simple as we show in the next session"
    - "It needs to inherit torch.nn.Module and just needs two methods: __init__ (so it’s an object) and forward (so it can take an image)."
        - I would explicitly say here "In plenoptic, all models are implemented using the library PyTorch. This means that every model must subclass the `torch.nn.Module` class and implement `forward`"
            - At the risk of sounding even more pedantic then usual, I think you don't strictly need an `__init__` method -- if you write `class MyModel(torch.nn.Module)` then you will already have `Module`'s `__init__` method
- As per API notes above, I would suggest moving `plenoptic.simulate.canonical_computations.filters` to `plenoptic.filters` or even to `plenoptic` if it is frequently used. If you don't want to move code around you could just add a `filters` package and import there from the lengthy namespace
- explain in models why inputs are 4d, like you do in conceptual -- define the dimensions
- suggest adding headings throughout quickstart
    - "Stimuli"
    - "Models"
    - "Synthesis"
        - state briefly here how we optimize / compute loss, since this is key concept. Link to relevant papers + additional tutorial/how-to
    - "Turn off gradient"
        - If this function is used frequently why not have it at package level? `plenoptic.remove_grad`
        - Is it absolutely necessary to `detach` all parameters in the model? Doesn't this limit the ability to use the library with arbitrary models, because a user would be surprised to find that some model they trained and wanted to continue training suddenly had all the parameters detached? Would it be sufficient to just set requires_grad of the parameters to False, e.g. by calling `Model.eval()`? Or what about using a context manager inside the Metamer class that sets requires_grad of model parameters to false just during the optimization step?
    - "Metamers"
    - "Eigendistortions"
- It was really unclear to me that we start with a randomly-initialized image. The part of the quickstart that uses the Marie Curie image makes this clearer, but I would say it explicitly in the first part
    - I would also explicitly say something like "at every step of optimization we compare the image, that begins as randomly initialized, with a representation provided by the model, and we use the loss function to minimize the difference between the representation of the metamer and the representation of the reference image
