### 2024-01-19 #150 @DanielaPamplona
Hi,

Sorry I got a cold so I could not finish on time. If I don't get worse, I expect to finish this on Monday.

Best

----------------------------------------------------------------------------------------------
Daniela Pamplona
Researcher and Lecturer at Université Paris-Est Créteil Val de Marne
https://danielapamplona.github.io/

Sent: Friday, January 19, 2024 at 7:51 PM
From: "Jonny Saunders" ***@***.***>
To: "pyOpenSci/software-submission" ***@***.***>
Cc: "Daniela Pamplona" ***@***.***>, "Mention" ***@***.***>
Subject: Re: [pyOpenSci/software-submission] Plenoptic (Issue #150)

Excellent, thank you David! @DanielaPamplona checking in if I can be of any assistance, or if you have a rough estimate of timeline? No rush! It looks like the authors have a few things to work on in the meantime :)
