### 2024-01-23 plenoptic-org/plenoptic#242 @DanielaPamplona
problem when installing with jupyter notebook 
I created a new env. in conda, installed plenoptic and jupyter notebook. When I tried to launch the notebook I got an error and it did not launch. I repeated the process without installing plenoptic and it worked.

Error below (copy-paste):

------------------------------
(p2) daniela@Vietnam:~$ jupyter notebook
Traceback (most recent call last):
  File "/home/daniela/anaconda3/envs/p2/lib/python3.9/site-packages/requests/compat.py", line 11, in <module>
    import chardet
ModuleNotFoundError: No module named 'chardet'

Traceback (most recent call last):
  File "/home/daniela/anaconda3/envs/p2/bin/jupyter-notebook", line 7, in <module>
    from notebook.app import main
  File "/home/daniela/anaconda3/envs/p2/lib/python3.9/site-packages/notebook/app.py", line 20, in <module>
    from jupyterlab.commands import (  # type:ignore[import-untyped]
  File "/home/daniela/anaconda3/envs/p2/lib/python3.9/site-packages/jupyterlab/__init__.py", line 8, in <module>
    from .handlers.announcements import (
  File "/home/daniela/anaconda3/envs/p2/lib/python3.9/site-packages/jupyterlab/handlers/announcements.py", line 15, in <module>
    from jupyterlab_server.translation_utils import translator
  File "/home/daniela/anaconda3/envs/p2/lib/python3.9/site-packages/jupyterlab_server/__init__.py", line 6, in <module>
    from .app import LabServerApp
  File "/home/daniela/anaconda3/envs/p2/lib/python3.9/site-packages/jupyterlab_server/app.py", line 15, in <module>
    from .handlers import LabConfig, add_handlers
  File "/home/daniela/anaconda3/envs/p2/lib/python3.9/site-packages/jupyterlab_server/handlers.py", line 21, in <module>
    from .listings_handler import ListingsHandler, fetch_listings
  File "/home/daniela/anaconda3/envs/p2/lib/python3.9/site-packages/jupyterlab_server/listings_handler.py", line 10, in <module>
    import requests
  File "/home/daniela/anaconda3/envs/p2/lib/python3.9/site-packages/requests/__init__.py", line 45, in <module>
    from .exceptions import RequestsDependencyWarning
  File "/home/daniela/anaconda3/envs/p2/lib/python3.9/site-packages/requests/exceptions.py", line 9, in <module>
    from .compat import JSONDecodeError as CompatJSONDecodeError
  File "/home/daniela/anaconda3/envs/p2/lib/python3.9/site-packages/requests/compat.py", line 13, in <module>
    import charset_normalizer as chardet
  File "/home/daniela/anaconda3/envs/p2/lib/python3.9/site-packages/charset_normalizer/__init__.py", line 23, in <module>
    from charset_normalizer.api import from_fp, from_path, from_bytes, normalize
  File "/home/daniela/anaconda3/envs/p2/lib/python3.9/site-packages/charset_normalizer/api.py", line 10, in <module>
    from charset_normalizer.md import mess_ratio
  File "charset_normalizer/md.py", line 5, in <module>
ImportError: cannot import name 'COMMON_SAFE_ASCII_CHARACTERS' from 'charset_normalizer.constant' (/home/daniela/anaconda3/envs/p2/lib/python3.9/site-packages/charset_normalizer/constant.py)

-------------------------------------------------------------------------

**System (please complete the following information):**
 - Ubuntu
 - Python version (as in the installation guide)
 - Pytorch version (as in the installation guide)
 - Plenoptic version (as in the installation guide)
