### 2024-01-23 LabForComputationalVision/plenoptic#242 @billbrod
Huh, I have never seen that kind of error. I can't reproduce this on my (Fedora) machine using `micromamba`, where I did the following:

```
micromamba create -n plen_nb 'python==3.9' -c conda-forge
# install from local copy
cd plenoptic/
pip install .
micromamba install -c conda-forge jupyterlab ipywidgets torchvision
jupyter lab
```

and that works. My understanding is that jupyter is encouraging everyone to go through `jupyter lab` now, not `notebook`, but running `micromamba install -c conda-forge notebook` after the above and then `jupyter notebook` works as well.

I'm not quite sure how to best debug this -- I'm not sure why there would be a `chardet` problem, which looks like it's a `requests` dependency, so maybe there's an incompatibility in the `requests` versions that end up getting used between the two? If so, then this is probably what I get for encouraging people to use conda for jupyter and pip for plenoptic, as you and @NickleDave pointed out.

A couple of possible ideas:
- Can you paste the exact sequence of commands you ran checking the version of `requests` after each of the install commands?
- If you try installing jupyterlab rather than notebook, as I did above, does that work?
- If you install jupyter (try both notebook and lab) and then plenoptic, do you get the same issues?
