### 2024-01-23 #150 @billbrod
Thanks Daniela! Like I said, it'll be a bit before i can dedicate time to most of these, but I wanted to try and solve the [jupyter issue](https://github.com/LabForComputationalVision/plenoptic/issues/242). However, I can't reproduce the issue and I'm not sure how to fix this -- @NickleDave or @sneakers-the-rat do you have any ideas?
