### 2024-01-24 LabForComputationalVision/plenoptic#242 @DanielaPamplona
Hi,
The commands were:

--------------------------------
conda create --name p2 pip python=3.9
conda activate p2
pip install plenoptic
conda install jupyter notebook
jupyter notebook
--------------------------------

If jupyter encourages to use lab instead of notebook, then I think that you can just write a statement in the installation saying that it *must* be installed lab. However, this solution must be discussed with the editor.

I also think that this might result from mixing pip and conda.
