### 2024-01-24 LabForComputationalVision/plenoptic#242 @billbrod
Okay, I'm still unable to reproduce this error following these steps. I'm using conda-forge as the only channel, so maybe that's the difference?

If you have time and can check the bullet points I wrote above, as well as what channels you're conda is using, that might help us get more specific.

But I think for now we assume it's a mixing of pip/conda issue (see related #104). I don't think it *won't* work with notebook, because I'm able to get it working with `notebook` locally. but lab might make it easier to get a working environment(?). Not really sure.
