### 2024-01-25 LabForComputationalVision/plenoptic#242 @DanielaPamplona
So, I checked and if I install first the notebook and then plenoptic there is no problem. There is also no problem when installing with jupyter lab. So, I would say, we just check this again after solving the pip issue.
