### 2024-01-26 LabForComputationalVision/plenoptic#242 @sneakers-the-rat
I was able to reproduce the issue (on a mac, if that matters). It's also not specific to jupyter (or plenoptic, for that matter) - if one tries to import `requests` after doing [this](https://github.com/LabForComputationalVision/plenoptic/issues/242#issuecomment-1908293499), you get the same error.

This looks like a `requests` problem initially:
https://github.com/psf/requests/issues/6384

but it turns out to be a jupyter problem.

requests tries to import either `chardet` or `charset_normalizer`:
https://github.com/psf/requests/blob/147c8511ddbfa5e8f71bbf5c18ede0c4ceb3bba4/requests/compat.py#L10-L15

they require `>=2,<4`: https://github.com/psf/requests/blob/147c8511ddbfa5e8f71bbf5c18ede0c4ceb3bba4/setup.cfg#L8

but the version that is installed is `2.0.4` - notably older than the current `3.3.2`.

Inspecting the `jupyter` and `notebook` conda packages, they both pin `charset-normalizer 2.0.4` (download the .tar.gz's from conda (anaconda), uncompress, see info > about.json).

uninstalling requests and reinstalling jupyter or notebook from pip resolved the problem.

It also seems to be just a problem with the `jupyter` and `notebook` packages in the default anaconda package source. using conda-forge also resolves the problem (`conda install -c conda-forge jupyter notebook`)

So I think adding the `-c conda-forge` resolves this
