### 2024-02-14 LabForComputationalVision/plenoptic#244 @billbrod
Update dependency descriptions
Making issue from [pyOpenSci review](https://github.com/pyOpenSci/software-submission/issues/150):

> I think it would be worth (1) stating clearly the dependency on torch, in the readme and in the installation page (see below) and (2) linking related libraries that are dependencies, e.g. https://github.com/LabForComputationalVision/pyrtools
