### 2024-02-14 LabForComputationalVision/plenoptic#245 @billbrod
Fix Zenodo badge on readme
Making issue from [pyOpenSci review](https://github.com/pyOpenSci/software-submission/issues/150):

> In the readme badge section, the link to Zenodo is just text that says DOI; why not use the Zenodo badge?

That's a bug, just needs to be fixed. (I swear it used to be a zenodo badge)
