### 2024-02-14 LabForComputationalVision/plenoptic#246 @billbrod
Improve API
From [pyOpenSci review](https://github.com/pyOpenSci/software-submission/issues/150):

> - In general, accessing core classes and functions is verbose. I have been guilty of this too and I know this would in a sense be a breaking change, but I would strongly suggest importing core classes and functions at the top-level if possible. I want to be able to just type po.Metamer and po.remove_grad. It could be fine to leave them where they are but just import them
> - Remember [file structure is an impelmentation detail and flat is better than nested](https://benhoyt.com/writings/python-api-design/#module-and-package-structure)
> - Reasonable people can disagree about this but [I prefer to avoid namespaces with names like tools or util](http://blog.nicholdav.info/four-tips-structuring-research-python/#finally-dont-name-a-module-packageutils-or-packageutilities), because [everything can go there](https://breadcrumbscollector.tech/stop-naming-your-python-modules-utils/), and as it stands now, you are just forcing users to type tools a lot. I would suggest moving modules in tools up a level and import the crucial functions directly into the plenopitc namespace from those modules. E.g. plenoptic.remove_grad. Or maybe plenoptic.grad.remove if you must have a module for it (seems weird that this function lives in validate).
> - it's also confusing and non-standard to internally import a sub-package at the root package level and change its name: e.g. synthesize -> synth and simulate -> simul. I kept thinking I missed an import statement with aliases, and I wasn't sure if there was something else going on with the code organization that I didn't know about. I would strongly suggest removing these internal aliases. If they are just too long, rename the module itself.
