### 2024-02-14 LabForComputationalVision/plenoptic#247 @billbrod
Packaging updates
Copying from [pyopensci review](https://github.com/pyOpenSci/software-submission/issues/150):

> I would change trove classifiers
> - the package is not in beta annymore if you are using in the lab
> - you can also specify specific pythons: 3.7, etc
>
> You probably want to raise the minimum python requires and adhere to [SPEC0](https://scientific-python.org/specs/spec-0000/) since you are depending on core scientific python libraries
>
> right now core packages are at 3.10-3.12, but if you need numba then you might need to raise it to 3.9 and test on 3.9-3.11 since https://github.com/numba/numba/issues/9197
