### 2024-02-20 06bb4a1
adds description of torch/pyrtools dependencies
