### 2024-02-20 7b3cbe8
updates citation discussion
