### 2024-02-20 LabForComputationalVision/plenoptic#240 @billbrod
I did not add something saying "If you are not using the code directly then cite the paper" -- I'm not sure what the case for that is. If someone isn't using plenoptic, they shouldn't cite our paper. If they're using the Portilla-Simoncelli texture, for example, without using plenoptic (because they're using the [matlab code](https://github.com/LabForComputationalVision/textureSynth)), they should just cite the relevant papers as given in the citation guide. (and I'm not sure why they'd be looking at our docs then!)
