### 2024-02-20 LabForComputationalVision/plenoptic#238 @billbrod
Alright, based on testing with @BalzaniEdoardo 's Mac, I'm 99% sure this is a result of getting a different sample of white noise during initialization.

My proposed fix is to *actually check against the stop criterion*, which makes more sense anyway.
