### 2024-02-20 fbd4f2b
fix for test_stop_criterion
on other OSs, can get a different sample of uniform noise when setting
the seed to 0, which means it doesn't always hit the stop criterion at
the same time. however, it should still *hit* the stop criterion, so we
just test against that directly
