+++ tests/test_steerable_pyr.py
@@ -81,7 +81,7 @@ class TestSteerablePyramid(object):
                                            for shape in [None, 224, '128_1', '128_2']])
     def img(self, request):
         im, shape = request.param.split('-')
-        img = po.load_images(op.join(DATA_DIR, f'256/{im}.pgm')).to(DEVICE)
+        img = po.load_images(DATA_DIR / "256" / f'{im}.pgm').to(DEVICE)
         if shape == '224':
             img = img[..., :224, :224]
         elif shape == '128_1':
@@ -93,7 +93,8 @@ class TestSteerablePyramid(object):
     @pytest.fixture(scope='class', params=[f'{shape}' for shape in [None, 224, '128_1', '128_2' ]])
     def multichannel_img(self, request):
         shape = request.param
-        img = po.load_images(op.join(DATA_DIR, f'512/flowers.jpg'), as_gray=False).to(DEVICE)
+        # use fixture for img and use color_wheel instead.
+        img = po.load_images(DATA_DIR / "mixed" / 'flowers.jpg', as_gray=False).to(DEVICE)
         if shape == '224':
             img = img[..., :224, :224]
         elif shape == '128_1':
