+++ src/plenoptic/data/__init__.py
@@ -1,0 +1,1 @@
+from .data_utils import get
