### 2024-02-20 f3a0cc2
fixed list get
