### 2024-02-20 d218e2f
moves osf_download to data_utils.py
so doesn't require users to have tests
