+++ tests/test_metric.py
@@ -1,114 +1,47 @@
 import pytest
-import requests
-import math
-import tqdm
-import tarfile
 import os
-import os.path as op
 import scipy.io as sio
 import torch
 import numpy as np
 import plenoptic as po
-from conftest import DATA_DIR, DEVICE
-
-
-# If you add anything here, remember to update the docstring in osf_download!
-OSF_URL = {'plenoptic-test-files.tar.gz': 'q9kn8', 'ssim_images.tar.gz': 'j65tw',
-           'ssim_analysis.mat': 'ndtc7', 'msssim_images.tar.gz': '5fuba', 'MAD_results.tar.gz': 'jwcsr',
-           'portilla_simoncelli_matlab_test_vectors.tar.gz': 'qtn5y',
-           'portilla_simoncelli_test_vectors.tar.gz': '8r2gq',
-           'portilla_simoncelli_images.tar.gz':'eqr3t',
-           'portilla_simoncelli_synthesize.npz': 'a7p9r',
-           'portilla_simoncelli_synthesize_torch_v1.12.0.npz': 'gbv8e',
-           'portilla_simoncelli_synthesize_gpu.npz': 'tn4y8',
-           'portilla_simoncelli_scales.npz': 'xhwv3'}
-
-
-def osf_download(filename):
-    r"""Download file from plenoptic OSF page.
-
-    From the OSF project at https://osf.io/ts37w/.
-
-    Downloads the specified file to `plenoptic/data`, extracts and deletes the
-    the .tar.gz file (if applicable), and returns the path.
-
-    Parameters
-    ----------
-    filename : {'plenoptic-test-files.tar.gz', 'ssim_images.tar.gz',
-                'ssim_analysis.mat', 'msssim_images.tar.gz',
-                'MAD_results.tar.gz',
-                'portilla_simoncelli_images.tar.gz',
-                'portilla_simoncelli_matlab_test_vectors.tar.gz',
-                'portilla_simoncelli_test_vectors.tar.gz',
-                'portilla_simoncelli_synthesize.npz',
-                'portilla_simoncelli_synthesize_torch_v1.12.0.npz',
-                'portilla_simoncelli_synthesize_gpu.npz'}
-        Which file to download.
-
-    Returns
-    -------
-    path : str
-        The path to the downloaded directory or file.
-
-    """
-    path = op.join(op.dirname(op.realpath(__file__)), '..', 'data', filename)
-    if not op.exists(path.replace('.tar.gz', '')):
-        print(f"{filename} not found, downloading now...")
-        # Streaming, so we can iterate over the response.
-        r = requests.get(f"https://osf.io/{OSF_URL[filename]}/download",
-                         stream=True)
-
-        # Total size in bytes.
-        total_size = int(r.headers.get('content-length', 0))
-        block_size = 1024*1024
-        wrote = 0
-        with open(path, 'wb') as f:
-            for data in tqdm.tqdm(r.iter_content(block_size), unit='MB',
-                                  unit_scale=True,
-                                  total=math.ceil(total_size//block_size)):
-                wrote += len(data)
-                f.write(data)
-        if total_size != 0 and wrote != total_size:
-            raise Exception(f"Error downloading {filename}!")
-        if filename.endswith('.tar.gz'):
-            with tarfile.open(path) as f:
-                f.extractall(op.dirname(path))
-            os.remove(path)
-        print("DONE")
-    return path.replace('.tar.gz', '')
+from conftest import DATA_DIR, DEVICE, IMG_DIR
 
 
 @pytest.fixture()
 def test_files_dir():
-    return osf_download('plenoptic-test-files.tar.gz')
[…]
