+++ src/plenoptic/data/__init__.py
@@ -1,1 +1,1 @@
-from .data_utils import get
+from .data_utils import get, download, osf_download
