### 2024-02-21 59e3e3e
removed type hinting for compatibility with python 3.7 and 3.8
