### 2024-02-21 c2b61a6
fix resources issue for python 3.8, 3.7
