+++ .github/workflows/ci.yml
@@ -86,15 +86,15 @@ jobs:
         python-version: ${{ matrix.python-version }}
         cache: pip
         cache-dependency-path: setup.py
+    - name: Create dummy duplicate (for test of plenoptic.data)
+      run: |
+        touch src/plenoptic/data/xyz.abc src/plenoptic/data/xyz.abcd
     - name: Install dependencies
       run: |
         # using the --upgrade and --upgrade-strategy eager flags ensures that
         # pip will always install the latest allowed version of all
         # dependencies, to make sure the cache doesn't go stale
         pip install --upgrade --upgrade-strategy eager .[dev]
-    - name: Create dummy duplicate (for test of plenoptic.data)
-      run: |
-        touch src/plenoptic/data/xyz.abc src/plenoptic/data/xyz.abcd
     - name: Run tests with pytest
       run: |
         pytest -n auto --cov-report xml
