### 2024-02-21 LabForComputationalVision/plenoptic#238 @NickleDave
Sorry for the slow reply

Not sure I follow all the logic of the test but I agree actually checking against the stop criterion makes more sense
