### 2024-02-21 LabForComputationalVision/plenoptic#240 @NickleDave
> I'm not sure what the case for that is. If someone isn't using plenoptic, they shouldn't cite our paper.

What about someone who's talking about the paper but isn't using the software?
If I'm using software, I cite the software.
But I get that the real world is what it is, and you want people to cite you.

Ignore the long thing I wrote above.
Here's what I think you should do:
- [ ] include whatever citation you want in the README, don't make someone go to the docs. So in other words, *do* repeat yourself in this case
- [ ] Also refer to the widget GitHub gives you from the citation.cff, since it's there anyways ("to copy-paste the bibtext click on the widget on the right")
- [ ] Finally, link to the list of all the citations in the docs; "for all related work, see [here](link)". You still want the citation for plenoptic at the top of that page for people that wind up there instead of the README
