+++ tests/test_data_get.py
@@ -1,5 +1,6 @@
 from contextlib import nullcontext as does_not_raise
 
+from importlib.abc import Traversable
 import pytest
 from torch import Tensor
 
@@ -20,10 +21,10 @@ def test_data_get_path(item_name, expectation):
         po.data.data_utils.get_path(item_name)
 
 
-# @pytest.mark.parametrize("item_name", ["color_wheel", "flowers", "curie"])
-# def test_data_get_path_type(item_name):
-#     """Test that the returned path object is an instance of Traversable."""
-#     assert isinstance(po.data.data_utils.get_path(item_name), Traversable)
+@pytest.mark.parametrize("item_name", ["color_wheel", "flowers", "curie"])
+def test_data_get_path_type(item_name):
+    """Test that the returned path object is an instance of Traversable."""
+    assert isinstance(po.data.data_utils.get_path(item_name), Traversable)
 
 
 @pytest.mark.parametrize(
