+++ tests/test_data_get.py
@@ -1,5 +1,6 @@
 from contextlib import nullcontext as does_not_raise
 
+from importlib import resources
 from importlib.abc import Traversable
 import pytest
 from torch import Tensor
@@ -17,8 +18,23 @@ import plenoptic as po
 )
 def test_data_get_path(item_name, expectation):
     """Test the retrieval of file paths with varying expectations."""
-    with expectation:
-        po.data.data_utils.get_path(item_name)
+    fh = resources.files("plenoptic.data")
+    # ensures that the files gets deleted, even if an exception gets hit in
+    # the try block.
+    try:
+        with expectation:
+            # Create files with specific extensions
+            for ext in ["abc", "abcd"]:
+                file_path = fh / f"xyz.{ext}"
+                file_path.touch(exist_ok=True)
+            po.data.data_utils.get_path(item_name)
+
+    finally:
+        # Delete the files created for the test
+        for ext in ["abc", "abcd"]:
+            path = fh / f"xyz.{ext}"
+            if path.exists():
+                path.unlink()  # This deletes the file
 
 
 @pytest.mark.parametrize("item_name", ["color_wheel", "flowers", "curie"])
