### 2024-02-21 8d8efa8
fix Demo Eigendistortion
data.get now can accept as_gray
