+++ README.md
@@ -18,8 +18,8 @@ which can be mapped to neuronal firing rate, fMRI BOLD response, behavior on
 some task, image category, etc. The intended audience is researchers in
 neuroscience, psychology, and machine learning. The generated stimuli enable
 interpretation of model properties through examination of features that are
-enhanced, suppressed, or descarded. More importantly, they can facilitate the
-scientific proceess, through use in further perceptual or neural experiments
+enhanced, suppressed, or discarded. More importantly, they can facilitate the
+scientific process, through use in further perceptual or neural experiments
 aimed at validating or falsifying model predictions.
 
 ## Getting started
