+++ docs/index.rst
@@ -35,7 +35,7 @@ plenoptic
    :align: center
    :alt: plenoptic logo
 
-``plenoptic`` is a python library for model-based synthesis of perceptual stimuli. For ``plenoptic``, models are those of visual [1]_ information processing: they accept an image as input, perform some computations, and return some output, which can be mapped to neuronal firing rate, fMRI BOLD response, behavior on some task, image category, etc. The intended audience is researchers in neuroscience, psychology, and machine learning. The generated stimuli enable interpretation of model properties through examination of features that are enhanced, suppressed, or descarded. More importantly, they can facilitate the scientific proceess, through use in further perceptual or neural experiments aimed at validating or falsifying model predictions.
+``plenoptic`` is a python library for model-based synthesis of perceptual stimuli. For ``plenoptic``, models are those of visual [1]_ information processing: they accept an image as input, perform some computations, and return some output, which can be mapped to neuronal firing rate, fMRI BOLD response, behavior on some task, image category, etc. The intended audience is researchers in neuroscience, psychology, and machine learning. The generated stimuli enable interpretation of model properties through examination of features that are enhanced, suppressed, or discarded. More importantly, they can facilitate the scientific process, through use in further perceptual or neural experiments aimed at validating or falsifying model predictions.
 
 
 Getting started
