### 2024-02-21 c58cd07
adds nbstripout before urlchecker
there are several cells in the notebook whose output includes links that
urlchecker htinks are invalid (though when I check them manually, they
look fine). Regardless, they're unimportant -- they're part of warning
messages and similar, so we don't want to check them.
