### 2024-02-21 77edd15
makes check_urls necessary.
since it shouldn't fail anymore
