### 2024-02-21 f0fea24
updates pyproject.toml
