### 2024-02-21 5b36541
updates python version to be compliant with SPEC0
