+++ jenkins/Dockerfile
@@ -2,12 +2,12 @@ FROM nvidia/cuda:12.2.0-devel-ubuntu20.04
 
 RUN apt-get update && \
     DEBIAN_FRONTEND=noninteractive apt-get install -y \
-      python3-numpy \
-      python3-pip \
-      python3-pytest \
-      python3-pytest-cov \
-      python3-venv \
-      python3-yaml \
+      python3.11-numpy \
+      python3.11-pip \
+      python3.11-pytest \
+      python3.11-pytest-cov \
+      python3.11-venv \
+      python3.11-yaml \
       ffmpeg \
       git \
       && \
