### 2024-02-21 LabForComputationalVision/plenoptic#240 @billbrod
I closed this before seeing your comment (by merging the PR), that's my bad.

I'm confused about
> >  I'm not sure what the case for that is. If someone isn't using plenoptic, they shouldn't cite our paper.
>
> What about someone who's talking about the paper but isn't using the software?
> If I'm using software, I cite the software.
> But I get that the real world is what it is, and you want people to cite you.

I think I misunderstood what you were saying -- when you said "the paper", you meant the plenoptic paper, for example: "let me discuss some awesome open-source projects in neuroscience, such as plenoptic [PLENOPTIC PAPER CITATION]". Is that right? I thought you had meant discussing the papers for the models/methods *implemented in plenoptic*, such as metamer or the texture model (in which case they don't need to cite plenoptic!).

I do agree with that, so I'll make the small change to add that and to add the extra link to the related work page.
