### 2024-02-21 LabForComputationalVision/plenoptic#238 @billbrod
Can you check with `main` ? I just merged in #251 , which should solve this issue.
