### 2024-02-21 LabForComputationalVision/plenoptic#246 @billbrod
@nickledave, I'd like your input on a potential restructure here:

```
plenoptic
   Metamer
   MADCompetition
   Geodesic
   Eigendistortion
   imshow
   animshow
   load_images
   remove_grad
   models
      all the classes found in models/ directory
   model_components
      everything found in canonical_computations/ directory
   metric
      everything found in metric/ directory
   conv
   display
   external
   ...all the other modules found in tools/ directory
```

My concern with the above is I have 10 files found in `tools/` (and I need to rename them, rearrange some of their functions). Do you think this would be too cluttered?

Separately, I have some helper functions for visualizing the results of synthesis optimization, such as `plenoptic.simulate.metamer.plot_synthesis_status`. These are all synthesis-specific (which is why they live in those specific files), but they're way too verbose to access. Do you have any suggestions as to how they should be accessed? They're not methods of the classes, because I didn't want to clutter those up -- I wanted the methods to all be synthesis/optimization-related, whereas these visualization ones all meant to be used after running optimization.
