### 2024-02-21 58d7c2b
added ruff and nox.
