+++ noxfile.py
@@ -1,0 +1,14 @@
+import nox
+
+
+@nox.session(name="Run Tests", python=["3.10", "3.11"])
+def tests(session):
+    """Run the test suite."""
+    session.run("pytest")
+
+
+@nox.session(name="linter", python = ["3.10", "3.11"])
+def linters(session):
+    """Run linters"""
+    session.run("ruff", "check", "--ignore", "D")
+
