### 2024-02-21 720dcd4
config nox for the py3.12
