+++ jenkins/Dockerfile
@@ -2,7 +2,7 @@ FROM nvidia/cuda:12.2.0-devel-ubuntu22.04
 
 RUN apt-get update && \
     DEBIAN_FRONTEND=noninteractive apt-get install -y \
-      'python-3.11' \
+      'python3.11' \
       'python3.11-venv' \
       ffmpeg \
       git \
