### 2024-02-21 cdda670
add `nox`, `ruff` and a section on managing multiple python versions
