### 2024-02-22 LabForComputationalVision/plenoptic#240 @NickleDave
Yes, that's what I meant, probably my fault for not being clearer. Fix sounds good to me
