### 2024-02-22 LabForComputationalVision/plenoptic#246 @NickleDave
> Do you think this would be too cluttered?

This list is what would be imported into the top-level name space, inside `plenoptic.__init__.py`?
The list above doesn't seem too cluttered, if those are all the main functions

For the couple of times where you list "all the other modules", do you absolutely need to import those into the top-level name space? Maybe you could use [lazy loading](https://scientific-python.org/specs/spec-0001/)? (Not trying to add to your to-do list, just a thought)

> Separately, I have some helper functions for visualizing the results of synthesis optimization, such as plenoptic.simulate.metamer.plot_synthesis_status. Do you have any suggestions as to how they should be accessed?

Very subjective but I really like to have a top-level `plot` module and throw all the functions in there, so that it is concise. So, something like `plenoptic.plot.metamer_synthesis_status`. The refactor would be annoying but it would make calling them more concise. Exactly because you want to interactively look at results e.g. after you run synthesis so you want to avoid being verbose there.
