### 2024-02-22 LabForComputationalVision/plenoptic#246 @billbrod
I was planning on looking into lazy loading, @DanielaPamplona mentioned it as well. I need to look into exactly what it means -- from the user perspective, it's similar to importing them in to the top-level name space, right?

I think getting rid of tools makes sense, it is an odd grab-bag of unrelated functions. I need to think more deeply about how much work it requires, but at the minimum I will import the most-used functions at the top level.

For `plot`: I think that does make sense. I think I like having a flat `plot` module, there aren't *too* many plotting functions. Both metamer and mad competition have corresponding `synthesis_status` methods, and there are a couple other corresponding (but slightly different) plotting functions, so I think I would have a `metamer_synthesis_status` and `mad_synthesis_status`.
