### 2024-02-22 cf1b0df
run all notebooks in same job
I was spending a lot of time on overhead. this removes get_notebook,
which also requires changing the all_tutorials_in_docs action
