+++ .github/workflows/ci.yml
@@ -9,22 +9,11 @@ on:
 
 jobs:
   # based on https://slashgear.github.io/how-to-split-test-by-folder-with-github-action/
-  get_notebooks:
-    runs-on: ubuntu-latest
-    outputs:
-      notebook: ${{ steps.get-notebooks.outputs.nb }}
-    steps:
-    - uses: actions/checkout@v4
-    - id: get-notebooks
-      run: "echo \"nb=$(ls examples/*ipynb | jq -R -s -c 'split(\"\\n\")[:-1]')\"\
-        \ >> $GITHUB_OUTPUT\n"
   notebooks:
     runs-on: ubuntu-latest
-    needs: [get_notebooks]
     strategy:
       matrix:
         python-version: [3.7, 3.8, 3.9, '3.10']
-        notebook: ${{fromJson(needs.get_notebooks.outputs.notebook)}}
       fail-fast: false
     name: Execute notebooks
     steps:
@@ -38,39 +27,33 @@ jobs:
       uses: FedericoCarboni/setup-ffmpeg@v2
     - name: Install dependencies
       run: |
-        pip install --upgrade --upgrade-strategy eager .[dev,nb]
+        pip install --upgrade --upgrade-strategy eager .[dev,nb] papermill
     - name: Download TID2013 dataset
-      if: ${{ matrix.notebook == 'examples/04_Perceptual_distance.ipynb' }}
       run: |
         mkdir -p data
         wget https://osf.io/7nfkz/download -O ./data/tid2013.rar
         7z x ./data/tid2013.rar -o./data/tid2013/
-    - name: Run notebooks
-      if: ${{ !contains(fromJSON('["examples/Demo_Eigendistortion.ipynb", "examples/Metamer-Portilla-Simoncelli.ipynb"]'), matrix.notebook) }}
-      run: jupyter execute ${{ matrix.notebook }}.ipynb --kernel_name=python3
-    - name: Run notebooks
-      if: ${{ matrix.notebook == 'examples/Metamer-Portilla-Simoncelli.ipynb' }}
-      # this notebook takes much longer than the rest (if run to completion,
-      # ~2hr on a laptop). We use papermill's parameters to reduce the max
-      # number of steps for metamer synthesis here (we want to test that each
-      # cell runs, but we don't need synthesis to go to completion). also,
-      # download images required for notebook
+    - name: Download Portilla-Simoncelli data
       run: |
-        mkdir -p data
         wget https://osf.io/eqr3t/download -O ./data/portilla_simoncelli_images.tar.gz
         tar xfz ./data/portilla_simoncelli_images.tar.gz --directory=./data/
-        pip install --upgrade --upgrade-strategy eager papermill
-        papermill ${{ matrix.notebook }} ${{ matrix.notebook }}_output.ipynb -p short_synth_max_iter 10 -p long_synth_max_iter 10 -p longest_synth_max_iter 10 -k python3 --cwd examples/
     - name: Run notebooks
-      if: ${{ matrix.notebook == 'examples/Demo_Eigendistortion.ipynb' }}
-      # this notebook takes much longer than the rest (if run to completion,
-      # ~1hr on a laptop, more than 5 hours on the Github runners). We use
-      # papermill's parameters to reduce the max number of iters for
-      # eigendistortion synthesis here (we want to test that each cell runs,
-      # but we don't need synthesis to go to completion)
+      if: ${{ !contains(fromJSON('["examples/Demo_Eigendistortion.ipynb", "examples/Metamer-Portilla-Simoncelli.ipynb"]'), matrix.notebook) }}
       run: |
-        pip install --upgrade --upgrade-strategy eager papermill
[…]
