+++ .github/workflows/ci.yml
@@ -81,7 +81,7 @@ jobs:
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@7598e39340e1dff4d6ebf7cf07a5e8184bde67e7 # v4.0.1
       with:
-        token: ${{ secreets.CODECOV_TOKEN }}
+        token: ${{ secrets.CODECOV_TOKEN }}
   all_tutorials_in_docs:
     runs-on: ubuntu-latest
     name: Check that all tutorial notebooks are included in docs
