### 2024-02-26 1bf81ef
updates description of expand/shrink
