### 2024-02-27 9886d06
restructure data module
use pooch to download external data, removes unnecesary images from
src/plenoptic/data
