+++ src/plenoptic/data/data_utils.py
@@ -1,26 +1,9 @@
-import math
-import os
-import pathlib
-import tarfile
 from importlib import resources
 from importlib.abc import Traversable
 from typing import Union
 
-import requests
-import tqdm
 
 from ..tools.data import load_images
-
-# If you add anything here, remember to update the docstring in osf_download!
-OSF_URL = {'plenoptic-test-files.tar.gz': 'q9kn8', 'ssim_images.tar.gz': 'j65tw',
-           'ssim_analysis.mat': 'ndtc7', 'msssim_images.tar.gz': '5fuba', 'MAD_results.tar.gz': 'jwcsr',
-           'portilla_simoncelli_matlab_test_vectors.tar.gz': 'qtn5y',
-           'portilla_simoncelli_test_vectors.tar.gz': '8r2gq',
-           'portilla_simoncelli_images.tar.gz':'eqr3t',
-           'portilla_simoncelli_synthesize.npz': 'a7p9r',
-           'portilla_simoncelli_synthesize_torch_v1.12.0.npz': 'gbv8e',
-           'portilla_simoncelli_synthesize_gpu.npz': 'tn4y8',
-           'portilla_simoncelli_scales.npz': 'xhwv3'}
 
 
 def get_path(item_name: str) -> Traversable:
@@ -78,86 +61,3 @@ def get(*item_names: str, as_gray: Union[None, bool] = None):
     if as_gray is None:
         as_gray = all(path.suffix == ".pgm" for path in paths)
     return load_images(paths, as_gray=as_gray)
-
-
-def download(url: str, destination: Union[str, pathlib.Path]) -> pathlib.Path:
-    r"""Download file from url.
-
-    Downloads file found at `url` to `destination`, extracts and deletes the
-    the .tar.gz file (if applicable), and returns the path.
-
-    Parameters
-    ----------
-    url :
-        Url of the file to download
-    destination :
-        Where to download the file.
-
-    Returns
-    -------
-    destination :
-        The path to the downloaded directory or file.
-
-    """
-    destination = pathlib.Path(destination)
-    # Streaming, so we can iterate over the response.
-    r = requests.get(url, stream=True)
-
-    # Total size in bytes.
-    total_size = int(r.headers.get('content-length', 0))
-    block_size = 1024*1024
-    wrote = 0
-    with open(destination, 'wb') as f:
-        for data in tqdm.tqdm(r.iter_content(block_size), unit='MB',
-                              unit_scale=True,
-                              total=math.ceil(total_size//block_size)):
-            wrote += len(data)
-            f.write(data)
-    if total_size != 0 and wrote != total_size:
-        raise Exception(f"Error downloading {destination.name} from {url}!")
-    if destination.name.endswith('.tar.gz'):
-        with tarfile.open(destination) as f:
-            f.extractall(os.path.dirname(destination))
-        os.remove(destination)
-        # need to call with_suffix twice to remove .tar and .gz
-        destination = destination.with_suffix('').with_suffix('')
-    return destination
-
-
-def osf_download(filename: str, destination_dir: Union[str, pathlib.Path] = '.') -> pathlib.Path:
-    r"""Download file from plenoptic OSF page.
-
-    From the OSF project at https://osf.io/ts37w/.
-
-    Downloads the specified file, extracts and deletes the the .tar.gz file (if
-    applicable), and returns the path.
-
-    Parameters
-    ----------
-    filename : {'plenoptic-test-files.tar.gz', 'ssim_images.tar.gz',
-                'ssim_analysis.mat', 'msssim_images.tar.gz',
-                'MAD_results.tar.gz',
-                'portilla_simoncelli_images.tar.gz',
-                'portilla_simoncelli_matlab_test_vectors.tar.gz',
[…]
