### 2024-02-27 c11f44d
removes data/ dir from github, puts on OSF
these are images used in tests, which we'll now download with others
using pooch
