### 2024-02-27 d407c7f
changes how we download test files
to using fetch_data
