+++ tests/utils.py
@@ -28,7 +28,7 @@ def update_ps_synthesis_test_file():
 
     """
     # for now, we want to compare against the one that worked for pytorch version 1.11
-    ps_synth_file = po.data.osf_download(get_portilla_simoncelli_synthesize_filename('1.11'))
+    ps_synth_file = po.data.fetch_data(get_portilla_simoncelli_synthesize_filename('1.11'))
     print(f'Loading from {ps_synth_file}')
 
     with np.load(ps_synth_file) as f:
