### 2024-02-27 c806648
updates how notebooks use po.data
