+++ src/plenoptic/data/__init__.py
@@ -2,7 +2,7 @@ from . import data_utils
 from .fetch import fetch_data, DOWNLOADABLE_FILES
 import torch
 
-__all__ = ['einstein', 'curie', 'Parrot', 'reptile_skin',
+__all__ = ['einstein', 'curie', 'parrot', 'reptile_skin',
            'fetch_data', 'DOWNLOADABLE_FILES']
 def __dir__():
     return __all__
@@ -16,8 +16,8 @@ def curie() -> torch.Tensor:
     return data_utils.get('curie')
 
 
-def Parrot(as_gray: bool = False) -> torch.Tensor:
-    return data_utils.get('Parrot', as_gray=as_gray)
+def parrot(as_gray: bool = False) -> torch.Tensor:
+    return data_utils.get('parrot', as_gray=as_gray)
 
 
 def reptile_skin() -> torch.Tensor:
