### 2024-02-28 5ea5653
updates notebooks for new data structure
- removes download from github ci (happens in notebook now)

- adds color_wheel back to src/data, since it's used in
  02_Eigendistortions

- adds tid2013.tar.gz to osf, and adds it to our fetch, for use in
  perceptual distance notebook

- change steerpyr notebook to using reptile_skin (from nuts)

- change imread to using data module
