### 2024-02-28 0fb1909
updates hash for test images
