### 2024-02-28 08e34fe
update test_images.tar.gz hash
