### 2024-02-28 63658de
set pooch allow_updates and retry_if_failed
