### 2024-02-28 442c19a
make sure pooch returns shared directory
