### 2024-02-28 98d6034
fix if statements in deploy
