+++ .github/workflows/deploy.yml
@@ -85,10 +85,10 @@ jobs:
           name: artifact
           path: dist
       - name: Publish package to test pypi
-        if: github.event_name == 'release' && github.event.action == 'published'
+        if: github.event_name == 'workflow_dispatch'
         uses: pypa/gh-action-pypi-publish@release/v1
         with:
           repository-url: https://test.pypi.org/legacy/
       - name: Publish package to real pypi
-        if: github.event_name == 'workflow_dispatch'
+        if: github.event_name == 'release' && github.event.action == 'published'
         uses: pypa/gh-action-pypi-publish@release/v1
