+++ .github/workflows/deploy.yml
@@ -1,5 +1,8 @@
 name: deploy
 on:
+  pull_request:
+    branches:
+    - main
   workflow_dispatch:
   release:
     types: [published]
@@ -85,7 +88,7 @@ jobs:
           name: artifact
           path: dist
       - name: Publish package to test pypi
-        if: github.event_name == 'workflow_dispatch'
+        if: github.event_name != 'release'
         uses: pypa/gh-action-pypi-publish@release/v1
         with:
           repository-url: https://test.pypi.org/legacy/
