### 2024-02-28 870fee5
sets fetch_depth: 0
otherwise don't get git tags correctly
