### 2024-02-28 2b94867
adds src to ruff config
