### 2024-02-28 4e3e1df
adds install ffmpeg (necessary to run notebooks)
