+++ .github/workflows/deploy.yml
@@ -1,8 +1,5 @@
 name: deploy
 on:
-  pull_request:
-    branches:
-    - main
   workflow_dispatch:
   release:
     types: [published]
