### 2024-02-28 d25dea2
updates ruff version (to work with src config)
