### 2024-02-28 LabForComputationalVision/plenoptic#235 @billbrod
Closed by #252
