### 2024-05-29 #150 @DanielaPamplona
Ahahah! Been there, no worries!

Get some sleep!

Daniela Pamplona

----------------------------------------------------------------------------------------------
Daniela Pamplona
Researcher and Lecturer at Université Paris-Est Créteil Val de Marne
https://danielapamplona.github.io/

Sent: Wednesday, May 29, 2024 at 5:29 PM
From: "William F. Broderick" ***@***.***>
To: "pyOpenSci/software-submission" ***@***.***>
Cc: "Daniela Pamplona" ***@***.***>, "Mention" ***@***.***>
Subject: Re: [pyOpenSci/software-submission] Plenoptic (Issue #150)

Hi all, just an update that I'm back from parental leave and will start working on this again, though my overall capacity is still reduced.
