### 2024-06-01 #150 @DanielaPamplona
Hi,

I am super proud of your strike and you have my full support.
If someone crosses the picket line, I will not do my part as well.

Stay stong!

Daniela

----------------------------------------------------------------------------------------------
Daniela Pamplona
Researcher and Lecturer at Université Paris-Est Créteil Val de Marne
https://danielapamplona.github.io/

Sent: Saturday, June 01, 2024 at 10:26 PM
From: "Jonny Saunders" ***@***.***>
To: "pyOpenSci/software-submission" ***@***.***>
Cc: "Daniela Pamplona" ***@***.***>, "Mention" ***@***.***>
Subject: Re: [pyOpenSci/software-submission] Plenoptic (Issue #150)

Fabulous!!! Hope you and the family are well. Just letting you know im on strike and will not be working until the strike is resolved. Whether someone else picking up the review is crossing a picket line is hard to say because this is volunteer work not for my employer, but I wont be proceeding with editorial duties until the University of California resolves its unfair labor practices with UAW 4811.
