### 2024-08-23 #150 @DanielaPamplona
Yes! That is my impression.

----------------------------------------------------------------------------------------------
Daniela Pamplona
Researcher and Lecturer at Université Paris-Est Créteil Val de Marne
https://danielapamplona.github.io/

Sent: Thursday, August 22, 2024 at 10:28 PM
From: "Jonny Saunders" ***@***.***>
To: "pyOpenSci/software-submission" ***@***.***>
Cc: "Daniela Pamplona" ***@***.***>, "Mention" ***@***.***>
Subject: Re: [pyOpenSci/software-submission] Plenoptic (Issue #150)

hi hello checking in here - i'm back at work and can continue editing this review, last i recall we were waiting on some responses to issues, correct?
