### 2024-08-23 #150 @billbrod
Hi, yes, I'm still working through the issues and reviews. I'll ping the reviewers on the two issues where I need them to try installing / running the tests again, to see if my changes fix the problems they have.

My next steps:
- I'm about to deprecate the `Geodesics` class, which isn't nearly as stable/robust as the rest of the library (I'm moving it to [another repo](https://github.com/plenoptic-org/geodesics) to be worked on separately). As part of that, I'm going to deprecate other functions that aren't used.
- Fixing several small bugs that I discovered when making recent changes.
- Change how my documentation is built, so that my example notebooks are run as part of the build, in order to make sure they stay up to date with any changes. (Currently, we run all the notebooks as part of the test, but don't run the optimization to completion, to save time, and then the documentation uses cached versions of the notebook that have been run to completion). This will involve moving over to [myst-nb](https://myst-nb.readthedocs.io/en/v0.13.2/index.html) and using myst more broadly.
- Change the documentation theme.

A lot of the feedback in the review was about structure / content of the documentation and the API, which I'll work on once I've done the restructuring above.

We are also [currently working on](https://github.com/plenoptic-org/plenoptic/pull/266) adding linters and making sure they pass

Does this sound like a reasonable plan to everyone? Should I post an intermediate check-in with what I've addressed so far before completing the above steps? Or wait until after I'm done with them?

(I've also moved the repo to a new organization, [plenoptic-org](https://github.com/plenoptic-org), so I just updated my initial comment to reflect that)
