### 2024-08-23 #150 @NickleDave
> Should I post an intermediate check-in with what I've addressed so far before completing the above steps? Or wait until after I'm done with them?

I don't think there's a need to do all that as long as we know you're converging. Understood you've got a lot on your plate right now!

It's up to @sneakers-the-rat obvs but sounds like a reasonable plan to me
