### 2025-02-05 #150 @billbrod
I'm really sorry that I've fallen behind on this. I am going to try and focus on it between now and [VSS](https://www.visionsciences.org/) this year (May). But otherwise same plan as before. A lot of the comments are documentation-related, and my plan is to switch to the pydata theme and myst, and then work through those issues, updating as I go.

Following up on my update above, I have deprecated geodesics and we now have linters running.

One additional change I made that I'm not sure if I posted about was to use Flatiron Institute's [jenkins](https://www.jenkins.io/) for the documentation build (instead of readthedocs), which allows us to use GPUs for building the docs, which I wanted to do before running the tutorial notebooks as part of the docs build (rather than using cached outputs), since a couple of them perform relatively heavy optimization.

I apologize again for the long delay on this and hope to have time to focus on it again this spring.
