### 2025-02-05 #150 @billbrod
I actually ran into an issue recently that I could use some input on: https://github.com/plenoptic-org/plenoptic/issues/313, based on a breaking change pytorch made for security reasons. Any input would be appreciated!
