### 2025-02-06 #150 @lwasser
@billbrod are you in our slack? you can ask questions there if you are and if not i'm happy to add you. It sounds like you could use a bit of help and that is what our community is best at!
