### 2025-02-07 #150 @billbrod
Hi @lwasser ! I am on the slack, but I'm not sure what belongs there versus here. Happy to post this question on the slack if you think it's relevant! Which channel makes sense?
