### 2025-02-20 #150 @JacksonBurns
Commenting from the slack!

Hi all, we have run into this on chemprop (machine learning for molecular properties) here: https://github.com/chemprop/chemprop/issues/1169

In our case, I think we save some scikit-learn scalers into our checkpoint files, which are now errorr-ing on torch 2.6

We haven't solved this yet, but our most likely solution is your 1st one (from here https://github.com/plenoptic-org/plenoptic/issues/313) - just 'revert' the change. Option 2 is OK but would require developers to continually add any new non-torch things to the safe list and any users extending the class to do the same. 3 is also OK but would require more work, and would also leave old model checkpoints non-functional, unless we wrote a conversion script.

Just overriding the default might be a security risk, but IMHO it's up to anyone deploying our software to ensure they aren't loading malicious checkpoints. That's a sharp edge we can't account for.
