### 2025-02-20 #150 @billbrod
Thanks Jackson! Before you posted, I was thinking that I would do option 3, but allow for users to pass `weights_only=False` themselves directly, so that our behavior would be safe moving forward, but users could override if they don't care and do trust their software. I need to think about how much work that would actually be, but I'm glad to hear how another scientific package is thinking about it.
