### 2025-02-28 #150 @billbrod
I have come up with a solution that I think works: https://github.com/plenoptic-org/plenoptic/pull/317. Any feedback would be appreciated!
