### 2025-03-04 #150 @billbrod
I am working on adding Examples sections to the user-facing functions (and running them as doctests) -- is there an easy way to get the percentage of user-facing functions with such a section? Like "doctest coverage"?
