### 2025-03-19 #150 @blink1073
There is!  https://interrogate.readthedocs.io/en/latest/index.html
