### 2025-03-19 #150 @blink1073
Ah, you're looking for specific content within the docstrings... you'd have to hook into something like numpydoc to do that I think...
