### 2025-03-20 #150 @Midnighter
I don't know about a coverage tool but pydoclint will complain about each public function that is not documented and you can configure what should be in the docstrings.
