### 2025-03-20 #150 @billbrod
Thanks all! I'm using ruff right now, which also has rules that complain about when public functions aren't documented (which I haven't turned on yet, but will). I'm not seeing something (in any of interrogate, pydoclint, or ruff), that will complain if an `Examples` section is missing.

However, I can make ruff complain if there's an empty section (like pydocstyle), and I could probably right a simple script to check whether everything has an Examples section (just with `"Examples" in obj.__doc__`). I think that will basically do what I want.

While it would be nice to get a report of actual doctest coverage, I'd settle for a CI action that fails if the percentage of public functions / methods / classes with an Examples section falls below some threshold, which I think I can do with the above.
