### 2025-03-20 #150 @agriyakhetarpal
Hi! I'm coming to this thread from Slack. The built-in [`sphinx.ext.coverage`](url) extension can calculate docstring coverage (works with Sphinx >=7). I had a reasonable amount of success with it when I used it a couple of years back, but it doesn't work well if you use `sphinx.ext.autodoc` ([this SO post is still valid in 2025](https://stackoverflow.com/q/33436025)).

Though, a solution already exists for this specific use case – `numpydoc.validate` will complain if an "Examples" section is missing in a docstring. Here's a list of the rules it can lint with: https://numpydoc.readthedocs.io/en/latest/validation.html#built-in-validation-checks

Alternatively, for a custom solution: you can simply `ast.walk()` over public classes and methods in the API with a custom Sphinx extension that connects to the `build-finished` event, or save a bit of effort and override `sphinx.ext.autodoc`'s behaviour by connecting to the  `autodoc-process-docstring` Sphinx event it provides. You can then raise warnings on whichever docstring doesn't have an "Examples" section by raising a warning (and optionally failing the build if you want it to, with `-W --keep-going`).
