### 2025-03-20 #150 @billbrod
Ah, that's excellent! Thanks for the pointers, I didn't realize that numpydoc has a validator built in! I'll look into adding that.
