### 2025-07-03 #150 @billbrod
Hi! I'm still working through this, and I had another question. As of https://github.com/plenoptic-org/plenoptic/pull/358, I'll have moved away from using jupyter notebooks to using myst-nb files, which sphinx builds into notebooks and are handled much better by git.

Currently, my git repo is really large because of the history of these (now removed) ipynb files (and there may have been some binary files we used to have, back when we just got started and didn't know what we were doing). In your opinion, is it worth trying to perform git surgery to remove these objects from the git history? (I could split out the ipynb files to a separate repo in case I need to look back through them for any reason.) My actual package, when installed from pypi or conda-forge, is much more reasonably sized, but the git repo is large. But messing with git history makes me nervous...

For scale: a fresh clone of the branch linked above (so, with the ipynb files removed) is 654MB, with 646MB of that in `.git`
