### 2025-10-07 #150 @lwasser
hey there @billbrod  and everyone on this review. We have undergone an editorial shift, and our previous and excellent editor, Jonny, is now focusing on other things around pyOpenSci after a wonderful terms with us. William, it appears that there is an unanswered question about Git history here. You can definitely discuss this with people in Slack if you wish. Notebooks will quickly bloat a repo. I'm a big fan of jupytext for that reason but myst-nb is similarly awesome. Do you still wish to proceed with this review? If you do, then we can find a new editor to move things forward. This one has been open for some time!!

I'm happy to help find a new editor for it to push things forward.
