### 2025-10-09 #150 @billbrod
Hi @lwasser thanks for pinging me about this again -- yes I would like to continue with this review. Sorry it's taking me so long!

For the git history question: I have switched over to using myst-nb (which is great!), but the problem is that because I previously had been using standard jupyter ipynb notebooks, my `.git` directory is huge. However, I am always leery about doing anything that involves editing the git history. So I'd like input on whether folks think it would be worth it to do git surgery to reduce the size of that directory, and, if so, any advice for how to best to do so.
