### 2025-12-16 #150 @eliotwrobson
@billbrod I would suggest reducing the size of that directory if possible, though that's really personal preference. Our review process isn't so concerned with details that granualar, so feel free to do whatever you think is best for managing this project.
