### 2026-01-15 #150 @lwasser
hi everyone! @eliotwrobson @billbrod I think we need to have a quick dm in Slack. 1. Jonny - the original editor has stepped down from being an editor during the time this is opened. This means we need to add a new editor to this review for it to move forward. But because the review has been open for so long, I want to ensure @billbrod that you are totally committed to pushing this forward. And I'd like to see if it happens in a few months, if possible. OR we can close it. Eliot is now our peer review lead, and he can help us move this forward. We need a commitment that you have the bandwidth to do so, and then we need a new editor to get it over the finish line! In 2023, I think this is our oldest open review. I'd love to see us resolve it one way or the other - whatever is best for all involved here, let's do that!
