### 2026-01-15 #150 @billbrod
Yes, I am interested in pushing this forward! I just messaged you and Eliot on slack.
