### 2026-03-05 #150 @billbrod
I believe I have now responded to all reviewers' comments. Sorry this has taken me so long and thanks especially to @NickleDave, @lwasser, and @eliotwrobson for helping me plan how to get this over the finish line.

Below, I will go through the comments, pointing out PRs, issues, or documentation as needed.

All changes required to respond to the reviews have been merged into main at this point, except for:
- adding Examples to every public function (see [issue 237](https://github.com/plenoptic-org/plenoptic/issues/237), which is tracking progress). In conversations with everyone, decided that, while I will still do this, it shouldn't hold up this review.
- the restructuring of the API / adding lazy-loading, which is in [PR 418](https://github.com/plenoptic-org/plenoptic/pull/418). That PR is still being reviewed by my team and will be merged after [PR 411](https://github.com/plenoptic-org/plenoptic/pull/411) (from an outside contributor).

My plan is:

- Merge 411
- Release 1.4.0
- Merge 418
- Address https://github.com/plenoptic-org/plenoptic/issues/417 (related to API restructure, but with @NickleDave, decided this shouldn't hold up this review)
- Release 2.0.0

Current docs found at https://docs.plenoptic.org/docs/branch/main, and for PR 418 (the API restructure) at https://docs.plenoptic.org/docs/pulls/418/. (Just going to https://docs.plenoptic.org will redirect to latest release, which doesn't include a lot of the documentation changes.)

Some additional questions I have:
- I would also like to do a JOSS submission for plenoptic (not sure why I didn't flag that originally). What is the process for that? Should we just finish up this review and then I inquire separately with JOSS?
- I used to have appropriate links to https://mybinder.org scattered across the docs. After the restructuring of my docs, they no longer work ([issue 409](https://github.com/plenoptic-org/plenoptic/issues/409) discusses what I would need to do to fix them again). However, it's not clear to me how many users actually find binder helpful. Do you have any ideas whether it's worth the effort / are they something pyOpenSci recommends? (Supporting it isn't hard, it's mostly the initial effort laid out in that issue.)

## Responding to @NickleDave's comments:

> Review Comments
> Required
>
> In this section I will put to-dos for items in the checklist above that are not checked\
>
>         Documentation
>             Examples for all user-facing functions DOC: Add examples for (most) user-facing functions [pyOpenSci review] plenoptic-org/plenoptic#237
>                 I find examples for some methods on main classes, e.g. plenoptic.synthesize.Metamer.load
>                 but many functions don't have examples, and some of these functions end up being core to the package, e.g. plenoptic.tools.remove_grad

I am still working on this (see [tracking issue](https://github.com/plenoptic-org/plenoptic/issues/237)). However, many functions and methods have had Examples added.

>         README
>             Badges for:
>                 A repostatus.org badge, DOC: Add repo status badge to README [pyOS review] plenoptic-org/plenoptic#236

Added in [PR 255](https://github.com/plenoptic-org/plenoptic/pull/255).

>             If applicable, how the package compares to other similar packages and/or how it relates to other packages in the scientific ecosystem.
>                 I think it would be worth (1) stating clearly the dependency on torch, here and in the installation page (see below) and (2) linking related libraries that are dependencies, e.g. https://github.com/LabForComputationalVision/pyrtools

I had added a paragraph about this (in [PR 248](https://github.com/plenoptic-org/plenoptic/pull/248/)), but just realized that I have since removed it (in [this commit](https://github.com/plenoptic-org/plenoptic/commit/b770b694567ba90a450b7466cf3be66eca0192da), part of [PR 410](https://github.com/plenoptic-org/plenoptic/pull/410/)), because I felt it was an odd mention: it was phrased as about the difficulty of installing pytorch or pyrtools on Windows, but I'm more confident in that now. After rereading David's comment I realize this probably wasn't what he was talking about, so I added some comments that I think are more inline with what he meant in [PR 421](https://github.com/plenoptic-org/plenoptic/pull/421) ([docs index](https://docs.plenoptic.org/docs/branch/main/#related-packages), [install page](https://docs.plenoptic.org/docs/branch/main/getting_started/install.html)).

Also, since our first discussion, I have come across [lucent](https://github.com/greentfrapp/lucent), which is a similar library (though it looks as if it's no longer maintained). I added a section linking to them as well in that same PR.

>         Usability
>
>         All functions have documentation and associated examples for use
>             this repeats the to-do from documentation above; I know it can be really tedious to write examples for literally every user-facing function but I would suggest that for frequently used functions such as plenoptic.tools.remove_grad there should be an example that passes when doctests are run

See point above about Examples for user-facing functions.

>         Functionality
>             Automated tests:
>                 All tests pass on the reviewer's local machine for the package version submitted by the author. Ideally this should be a tagged version making it easy for reviewers to install.
>                     this test failed for me locally: tests/test_metamers.py::TestMetamers::test_stop_criterion TST: Test fails locally [pyOpenSci review] plenoptic-org/plenoptic#238

Discussed in [issue 238](https://github.com/plenoptic-org/plenoptic/issues/238), closed in [PR 294](https://github.com/plenoptic-org/plenoptic/pull/294).

>         Packaging guidelines: The package conforms to the pyOpenSci packaging guidelines.
>             Code format is standard throughout package and follows PEP 8 guidelines (CI tests for linting pass)
>                 currently there's no code formatting or linting (e.g. flake8), AFAICT -- left a comment on an existing issue re: formatters here Consistent formatting (Python Black) plenoptic-org/plenoptic#51 (comment)

Ruff added in [PR 266](https://github.com/plenoptic-org/plenoptic/pull/266), modified further in [PR 362](https://github.com/plenoptic-org/plenoptic/pull/362), now runs as part of pre-commit hook (and with pre-commit CI on every PR).

> suggested
>
> In this section I will put suggested changes. Some of these start to get very nit-picky so I won't raise issues for all of them
>
>     documentation
>         be clear about what "model" means in the context of plenoptic: these are models of visual perception, although there are some models of audio perception as well. This should be the second sentence of explaining what plenoptic does, right after saying "it's for synthesizing stimuli using models" DOC: Clarify what "model" means in README + conceptual introduction [pyOpenSci review] plenoptic-org/plenoptic#239

Discussed in [issue 239](https://github.com/plenoptic-org/plenoptic/issues/239), closed in [PR 250](https://github.com/plenoptic-org/plenoptic/pull/250). This is now described at top of [docs index](https://docs.plenoptic.org/docs/branch/main/index.html) and [conceptual introduction](https://docs.plenoptic.org/docs/branch/main/getting_started/conceptual_intro.html)

>     README
>         Badge section
>             In the badge section, the link to Zenodo is just text that says DOI; why not use the Zenodo badge?
>                 DOI

That was an accident. Initially fixed in [PR 249](https://github.com/plenoptic-org/plenoptic/pull/249), then further updated in [PR 382](https://github.com/plenoptic-org/plenoptic/pull/382).

>         citation
>             There is a citation section but it just links to a list of citations in the docs. I would suggest the following
>                 Directly state in this section of the README how the code should be cited. E.g., "If you use the code, please cite both the code by DOI as well as the JOV paper. If you are not using the code directly then cite the paper". I would also take advantage of the citation.cff, immediately after say "you can also click on 'cite this repository' on the right side of the page to get a copyable citation". I would then include both the Zenodo badge with the code DOI and a bibtex citation for the paper in this section, just in case someone misses the part about "click on the right page". Finally I would say "for more in-depth citation lists go here (link to citation page)" DOC: clearly state how to cite in citation section of README and docs, point to citation.cff [pyOpenSci review] plenoptic-org/plenoptic#240

Discussed in [issue 240](https://github.com/plenoptic-org/plenoptic/issues/240), closed by [PR 249](https://github.com/plenoptic-org/plenoptic/pull/249) and [256](https://github.com/plenoptic-org/plenoptic/pull/256). You can see what this looks like in [readme](https://github.com/plenoptic-org/plenoptic?tab=readme-ov-file#citing-us) and [docs](https://docs.plenoptic.org/docs/branch/main/reference/citation.html#citation-doc) (docs index just points to dedicated page).

>     API
>         In general, accessing core classes and functions is verbose. I have been guilty of this too and I know this would in a sense be a breaking change, but I would strongly suggest importing core classes and functions at the top-level if possible. I want to be able to just type po.Metamer and po.remove_grad. It could be fine to leave them where they are but just import them
>         Remember file structure is an impelmentation detail and flat is better than nested
>         Reasonable people can disagree about this but I prefer to avoid namespaces with names like tools or util, because everything can go there, and as it stands now, you are just forcing users to type tools a lot. I would suggest moving modules in tools up a level and import the crucial functions directly into the plenopitc namespace from those modules. E.g. plenoptic.remove_grad. Or maybe plenoptic.grad.remove if you must have a module for it (seems weird that this function lives in validate).
>
>     it's also confusing and non-standard to internally import a sub-package at the root package level and change its name: e.g. synthesize -> synth and simulate -> simul. I kept thinking I missed an import statement with aliases, and I wasn't sure if there was something else going on with the code organization that I didn't know about. I would strongly suggest removing these internal aliases. If they are just too long, rename the module itself.

Thanks for your help coming up with a better API, these changes are discussed in [issue 246](https://github.com/plenoptic-org/plenoptic/issues/246) and addressed in  [PR 418](https://github.com/plenoptic-org/plenoptic/pull/418). As mentioned above, this PR has not been merged yet.

>     installation
>         I would not suggest installing with pip into a conda environment
>             because conda recommends only installing with pip last after everything else
>             would it be possible to make a conda-forge package?
>                 existing issue: Put on conda plenoptic-org/plenoptic#104
>                 my comment: Put on conda plenoptic-org/plenoptic#104 (comment)

We are now on conda-forge. see discussion in [issue 104](https://github.com/plenoptic-org/plenoptic/issues/104) and updated instructions in [PR 278](https://github.com/plenoptic-org/plenoptic/pull/278).

Install instructions have been updated to separate out conda/pip and help folks with the decision, see [current install docs](https://docs.plenoptic.org/docs/branch/main/getting_started/install.html). I am planning on updating these instructions to discuss using `uv` ([issue 400](https://github.com/plenoptic-org/plenoptic/issues/400)), since that's what I currently use.

>     documentation
>         consider switching to newer theme like pydata-sphinx?
>         As stated above, I would strongly suggest making it as obvious as possible that most functionality requires torch
>             in install instructions
>         Some thoughts on organization, as per the diataxis framework: Reorganize Documentation structure plenoptic-org/plenoptic#21 (comment), Reorganize Documentation structure plenoptic-org/plenoptic#21 (comment)
>             put install + quickstart in a "getting started"
>             put model requirements and citation guide in a "reference"
>             most everything else should go in a "user guide"
>             except API should have its own section
>             reproducibility should go in reference

Done in [PR 410](https://github.com/plenoptic-org/plenoptic/pull/410), see response to Daniela's review below for some more explanation of the organization I landed on.

I have also been thinking about adding more of what diataxis terms "How-to Guides". To me, those are like the Examples section in the docstring, the examples in [matplotlib](https://matplotlib.org/stable/gallery/index), and the gallery in [seaborn](https://seaborn.pydata.org/examples/index.html). I'd have to think about how to have such bite-sized examples for plenoptic (and how to give a good overview of them), but given my tendency for long-windedness, this is probably a good exercise. Any suggestions would be appreciated (as would examples of good "How-to Guides" from other packages, especially non-plotting ones).

>         API docs
>             formatting looks off here in some places -- I know that sphinx-autodoc is not the easiest thing to work with but some clean-up here would help it look more approachable
>             I added a typo on this issue: Sphinx & docs quality plenoptic-org/plenoptic#102 (comment), will add more there if I find them

Started in [PR 386](https://github.com/plenoptic-org/plenoptic/pull/386), more work in [PR 413](https://github.com/plenoptic-org/plenoptic/pull/413). API docs names are very long in main, are better after [PR 418](https://github.com/plenoptic-org/plenoptic/pull/418), docs for which can be seen [here](https://docs.plenoptic.org/docs/pulls/418/api/index.html).

>     functionality / presentation
>     I tested by running notebooks, mostly.
>     I kept notes for the key notebooks as I read through them. This starts to get quite nitpicky, please know I am trying to help with presentation of material. Going to add here for now but I could move this to individual issues.
>         quickstart
>             the function load_images in the notebook did not work for me at first
>             I read the function but it uses op -- I would suggest making the code as readable as possible by not abbreviating and used absolute names instead, esp for standard library functions; code is read much more than it's written
>             The problem was that the notebook expects to be run from inside notebook and so it has a hardcoded path
>                 a fix for this would be to run notebooks from root
>                 and/or to include images in the package as discussed in (link issue)
>             I would also suggest re-writing this with pathlib; pathlib.Path().exists()
>             and raising an explicit error like "file does not exist: {}"

Issues with `load_images` discussed in [issue 235](https://github.com/plenoptic-org/plenoptic/issues/235), fixed in [PR 252](https://github.com/plenoptic-org/plenoptic/pull/252). Similar use of `importlib` for other data files addressed in [PR 289](https://github.com/plenoptic-org/plenoptic/pull/289). Final replacement of `os.path` (and `op`) in [PR 414](https://github.com/plenoptic-org/plenoptic/pull/414)

>     packaging guidelines
>         pyproject.toml
>             I would change trove classifiers
>                 the package is not in beta annymore if you are using in the lab
>                 you can also specify specific pythons: 3.7, etc
>             You probably want to raise the minimum python requires and adhere to SPEC0 since you are depending on core scientific python libraries
>             right now core packages are at 3.10-3.12, but if you need numba then you might need to raise it to 3.9 and test on 3.9-3.11 since numba is not out for 3.12 just yet

Following your comments, I was trying to adhere to SPEC-0000 but have since decided that I will match pytorch's supported python versions, since that's my main dependency. I have added an action to track their supported versions and raise an error if my supported versions diverge from theirs. (Currently doing this by parsing the classifiers in `pyproject.toml`, because I couldn't figure out how to do this by querying `pip` directly, happy to update if there is a way to do so.)

Does that seem reasonable?

Switch to SPEC-0000 done in [PR 254](https://github.com/plenoptic-org/plenoptic/pull/254) and to pytorch python versions in [PR 423](https://github.com/plenoptic-org/plenoptic/pull/423).

> Reading notes
> conceptual introduction
>
>     ""> Synthesis is a framework for exploring models" -- what kind of models? (Answer: cognitive/neural!)
>
>     index
>
>     "descarded" -> "discarded"

Closed in [PR 250](https://github.com/plenoptic-org/plenoptic/pull/250).

> quickstart
>
>     "Models can be really simple, as this demonstrates."
>         I think you want to say something here like "Models can be really simple as we show in the next session"
>         "It needs to inherit torch.nn.Module and just needs two methods: init (so it’s an object) and forward (so it can take an image)."
>             I would explicitly say here "In plenoptic, all models are implemented using the library PyTorch. This means that every model must subclass the torch.nn.Module class and implement forward"
>                 At the risk of sounding even more pedantic then usual, I think you don't strictly need an __init__ method -- if you write class MyModel(torch.nn.Module) then you will already have Module's __init__ method
>     As per API notes above, I would suggest moving plenoptic.simulate.canonical_computations.filters to plenoptic.filters or even to plenoptic if it is frequently used. If you don't want to move code around you could just add a filters package and import there from the lengthy namespace
>     explain in models why inputs are 4d, like you do in conceptual -- define the dimensions
>     suggest adding headings throughout quickstart
>         "Stimuli"
>         "Models"
>         "Synthesis"
>             state briefly here how we optimize / compute loss, since this is key concept. Link to relevant papers + additional tutorial/how-to
>         "Turn off gradient"
>             If this function is used frequently why not have it at package level? plenoptic.remove_grad
>             Is it absolutely necessary to detach all parameters in the model? Doesn't this limit the ability to use the library with arbitrary models, because a user would be surprised to find that some model they trained and wanted to continue training suddenly had all the parameters detached? Would it be sufficient to just set requires_grad of the parameters to False, e.g. by calling Model.eval()? Or what about using a context manager inside the Metamer class that sets requires_grad of model parameters to false just during the optimization step?
>         "Metamers"
>         "Eigendistortions"
>     It was really unclear to me that we start with a randomly-initialized image. The part of the quickstart that uses the Marie Curie image makes this clearer, but I would say it explicitly in the first part
>         I would also explicitly say something like "at every step of optimization we compare the image, that begins as randomly initialized, with a representation provided by the model, and we use the loss function to minimize the difference between the representation of the metamer and the representation of the reference image

[…]
