### 2026-03-11 #150 @billbrod
Thanks Eliot and David! I'll add the badge and release after [PR 418](https://github.com/plenoptic-org/plenoptic/pull/418) is merged. I'm about to head to a conference (and the rest of my team is already there), so that won't happen right away, but I'll post here when it does. After we do that, I'll go ahead and submit to JOSS.

Thanks again for all your help!
